### 2025-11-12 #730 @kylesnap
ernest: A Toolkit for Nested Sampling
Submitting Author  @kylesnap
Repository:  https://github.com/kylesnap/ernest
Version submitted:1.1.0
Submission type: Stats
Badge grade: bronze

Version accepted: TBD
---

```
Package: ernest
Title: A Toolkit for Nested Sampling
Version: 1.1.0
    person("Kyle", "Dewsnap", , "kyle.dewsnap@ubc.ca", role = c("aut", "cre"),
Description: Bayesian evidence estimation and posterior inference with the
    nested sampling algorithm, along with S3 methods for simulating
    uncertainty and creating visualisations.
    cli,
    generics,
    glue,
    lifecycle,
    matrixStats,
    posterior,
    prettyunits,
    rlang (>= 1.1.0),
    uniformly,
    vctrs,
    withr
    ggdist,
    LaplacesDemon,
    log4r,
    tidyselect,
    truncnorm,
    vdiffr,
    xml2
    cpp11,
    cpp11eigen,
    knitr
Config/testthat/parallel: true
Config/testthat/start-first: algorithm, *_ellipsoid, mini_balls
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [X] A pre-submission inquiry has been approved in [issue#722](#722).

- The target audience includes statisticians and researchers who use Bayesian inference and require robust estimation of model evidence and posterior distributions. ernest can be used to support Bayesian model comparison and parameter estimation, and has been built to allow users to specify their own likelihood functions and prior space specifications depending on their discipline.

- This is the first package that implements a complete toolkit for performing nested sampling in R, including an implementation of the nested sampling algorithm. Certain nested sampling implementations built for other languages do offer interfaces for R (see README). However, ernest's use of S3 generics and compatibility with popular external packages (such as ggplot2 and Stan's posterior) aim to make this package more usable and performant than non-native options.

- What grade of badge are you aiming for? Bronze.

- [X] I have read the [*Statistical Software Peer Review* Guide for Authors](https://stats-devguide.ropensci.org/pkgdev.html).
