### 2025-12-03 #730 @andrewheiss
Yep, that's fine!
