### 2025-12-10 kylesnap/ernest#23 @kylesnap
Comparing ernest to LRPS.
Following the comments made by @saudiwin (https://github.com/ropensci/software-review/issues/730#issuecomment-3634114889), ernest would benefit from a comparison to an existing package like `brms`. I'll rework ernest's package vignette so that it includes a comparison against the worked example from the main `brms` vignette.
