### 2025-12-15 kylesnap/ernest#28 @kylesnap
Expected types and validation for likelihood and prior functions.
Related to https://github.com/ropensci/software-review/issues/730#issuecomment-3634114889, ernest's documentation and implementation of the `create_prior` and `create_likelihood` function factories should clearly describe the expected types of the input and outputs of user-entered functions.
