### 2026-01-09 kylesnap/ernest#29 @kylesnap
Address editor feedback.
List based on [editor feedback](https://github.com/ropensci/software-review/issues/730#issuecomment-3720857961).

## Documentation/Vignettes
- [x] Rewrite Bayes’ theorem explanation using explicit conditional probabilities appear consistently.
- [x] Resolve terminology inconsistencies in nested sampling vignettes (e.g., between shells vs cells).
- [x] Consider a glossary of key terms (call, iteration, point / live point)
- [x] Document how vectorized likelihoods work and their contribution to `n_call`
- [x] For summary(): Explain why summary(run)$run has iterations + n_points rows, and clearly document what the call column indexes and how it relates to iterations and likelihood evaluations.
- [x] Improve explanation of supplying data via likelihood functions (i.e., function(x) function(y) {...})
## Statistical Notes:
- [x] Fix HDCI: Either import a package or rename it to what it actually is (ETI).
- [x] Explain why resampling is required in `visualize` (i.e., more detail on posterior probability weights from NS).
## Code Quality & Design Improvements
- [x] Resolve loose TODOs.
- [x] Consider simplifying CLI styling for output.
- [x] Preserve or expose whether a sampler seed was user-supplied vs auto-generated.
- [x] Refactor stopping-criteria logic to avoid reliance on c(NULL, NULL, 0).
- [x] Refactor check_matrix() to use vectorized checks instead of row-wise loops.
- [x] Remove unnecessary @inheritParams stats::qnorm from create_normal_prior().
- [x] Fix incorrect documentation claiming use of make.unique() when vctrs::vec_as_names() is used.
## Testing Strategy & Reliability
- [x] Reduce reliance on snapshot tests.
- [x] Remove snapshot tests that capture progress bars, spinners, or transient CLI output.
