### 2026-02-05 #730 @kylesnap
@andrewheiss, I'm preparing a pull request with changes suggested by the editors (https://github.com/kylesnap/ernest/pull/31). Am I okay to merge these into the main branch now, or should I wait for additional review from you before doing so?
