### 2026-03-18 c7addbd
Cleaner evidence calculation.
