### 2026-03-18 36f8dd8
Reworked RNG seed handling.
