### 2026-03-19 907208b
Faster plotting methods.
