+++ R/plot.R
@@ -1,8 +1,8 @@
-#' Plot diagnostics for a nested sampling run
+#' Plot diagnostics from nested sampling results
 #'
-#' Visualizes key diagnostics from a nested sampling run, including the
-#' normalized likelihood, posterior weights, and evidence as functions
-#' of log-prior volume.
+#' Visualizes key diagnostics from nested sampling outputs, including the
+#' normalized likelihood, posterior weights, and evidence as functions of
+#' log-prior volume.
 #'
 #' @param x [[ernest_run]] or [[ernest_estimate]]\cr An object containing
 #' results from nested sampling.
@@ -14,9 +14,9 @@
 #' side effect.
 #'
 #' @details
-#' Interpreting these plots can help diagnose issues such as poor or
-#' insufficient prior sampling and model misspecification. Use `which` to select
-#' the plots to display:
+#' The plotting methods are structured around internal `autoplot_*` helpers
+#' (`autoplot_evidence()`, `autoplot_weight()`, and `autoplot_likelihood()`).
+#' Use `which` to select one or more diagnostics:
 #'
 #' * `which = "evidence"`: Plots the estimated marginal
 #' likelihood (evidence) as a function of log-prior volume, with uncertainty
@@ -32,18 +32,20 @@
 #' irregularities may suggest sampling difficulties or, in some cases,
 #' misspecified likelihood functions.
 #'
-#' If `x` is an `ernest_run`, the plots are based on the actual run data. Error
-#' ribbons are drawn around the `evidence` plot from analytic estimates of
-#' uncertainty (see {summary.ernest_run}).
+#' If `x` is an `ernest_run`, plotting uses deterministic integral estimates
+#' (`calculate(x, ndraws = 0)`). For `which = "evidence"`, uncertainty bands
+#' are derived from the analytic standard error stored in
+#' `x$log_evidence_err`.
 #'
 #' If `x` is an `ernest_estimate` (or if `ndraws` is specified), the plots
-#' are based on simulated values from the log-volume. The highest density
-#' continuous intervals (HDCIs) are computed using [ggdist::median_hdci()] for
-#' both the `evidence` and `weight` plots.
+#' are based on simulated log-volume trajectories. In this case,
+#' `plot.ernest_estimate()` requires `attr(x, "ndraws") > 1`. Uncertainty for
+#' `evidence` and `weight` is summarized using `ggdist` interval helpers
+#' (`point_interval()` and `curve_interval()`) over interpolated curves.
 #'
 #' @note
-#' Plotting multiple diagnostics with `which` requires the patchwork package.
-#' Plotting `ernest_estimate` objects requires the ggdist package.
+#' Plotting multiple diagnostics with `which` requires the {patchwork} package.
+#' Plotting `evidence` or `weight` diagnostics requires the {ggdist} package.
 #'
 #' @srrstats {BS6.1} Default plot for return object.
 #'
@@ -72,17 +74,7 @@ plot.ernest_estimate <- function(
   ...
 ) {
   check_dots_empty()
-  which <- arg_match(
-    which,
-    values = c("evidence", "weight", "likelihood"),
-    multiple = TRUE
-  )
-  if (length(which) == 0) {
-    cli::cli_abort("At least one plot type must be specified in `which`.")
-  } else if (length(which) > 0) {
-    check_required("patchwork", "to combine multiple plots")
-  }
[…]
