### 2026-03-19 fe8abd8
Reorganized helpers for plotting.
