+++ R/plot.R
@@ -116,7 +116,7 @@ check_plot_which <- function(which, call = caller_env()) {
       "At least one plot type must be specified in `which`.",
       call = call
     )
-  } else if (length(which) > 0) {
+  } else if (length(which) > 1) {
     check_required("patchwork", "to combine multiple plots", call = call)
   }
   which
@@ -164,9 +164,16 @@ autoplot.ernest_estimate <- function(object, which, call = caller_env(), ...) {
         log_volume,
         log_evidence,
         xout,
-        xintercept = xint
+        xintercept = xint,
+        call = call
       ),
-      "weight" = autoplot_weight(log_volume, weight, xout, xintercept = xint),
+      "weight" = autoplot_weight(
+        log_volume,
+        weight,
+        xout,
+        xintercept = xint,
+        call = call
+      ),
       "likelihood" = autoplot_likelihood(log_volume, lik, xintercept = xint)
     )
   })
@@ -210,9 +217,16 @@ autoplot.ernest_run <- function(object, which, call = caller_env(), ...) {
         log_evidence,
         xout,
         log_evidence_err = object$log_evidence_err,
-        xintercept = xint
+        xintercept = xint,
+        call = call
       ),
-      "weight" = autoplot_weight(log_volume, weight, xout, xintercept = xint),
+      "weight" = autoplot_weight(
+        log_volume,
+        weight,
+        xout,
+        xintercept = xint,
+        call = call
+      ),
       "likelihood" = autoplot_likelihood(log_volume, lik, xintercept = xint)
     )
   })
@@ -254,8 +268,10 @@ autoplot_evidence <- function(
   xout,
   nbands = 3,
   log_evidence_err = NULL,
-  xintercept = NULL
+  xintercept = NULL,
+  call = caller_env()
 ) {
+  check_installed("ggdist", "to plot evidence diagnostics", call = call)
   z_df <- if (!is.null(log_evidence_err)) {
     fill_name <- "\U00B1 \U03C3"
     df <- tibble::tibble(
@@ -274,29 +290,14 @@ autoplot_evidence <- function(
     df
   } else {
     fill_name <- "MHD"
-    nout <- length(xout)
-    yout <- posterior::rvar(
-      t(vapply(
-        seq_len(nrow(log_volume)),
-        \(i) {
-          approx(
-            log_volume[i, ],
-            exp(log_evidence[i, ]),
-            rule = 2,
-            xout = xout
-          )$y
-        },
-        double(nout)
-      )),
-      dim = c(nout)
+    yout <- interpolate_draw_curves(
+      nrow(log_volume),
+      length(xout),
+      \(i) {
+        approx(log_volume[i, ], exp(log_evidence[i, ]), rule = 2, xout = xout)$y
+      }
     )
-    tibble::tibble(
-      x = rep(xout, nbands),
-      !!!ggdist::point_interval(
-        yout,
-        .width = ggdist::interval_widths(nbands)
-      )
-    )
+    interval_df(xout, posterior::rvar(yout), nbands, type = "point")
   }
 
   ggplot(z_df, aes(.data[["x"]], y = .data[[".value"]])) +
@@ -334,10 +335,13 @@ autoplot_weight <- function(
   weight,
   xout,
   nbands = 3,
-  xintercept = NULL
+  xintercept = NULL,
+  call = caller_env()
 ) {
-  yout <- t(vapply(
-    seq_len(nrow(log_volume)),
+  check_installed("ggdist", "to plot importance weights", call = call)
+  yout <- interpolate_draw_curves(
+    nrow(log_volume),
+    length(xout),
     \(i) {
[…]
