+++ man/merge.ernest_run.Rd
@@ -1,0 +1,56 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/merge.R
+\name{merge.ernest_run}
+\alias{merge.ernest_run}
+\title{Merge two nested sampling runs}
+\usage{
+\method{merge}{ernest_run}(x, y, ...)
+}
+\arguments{
+\item{x, y}{[\link{ernest_run}]\cr
+Runs generated from compatible nested sampling specifications.}
+
+\item{...}{These dots are for future extensions and must be empty.}
+}
+\value{
+[\link{ernest_run}] containing merged dead points and live points.
+}
+\description{
+Combines two nested sampling runs into a single run by re-ordering their
+samples by log-likelihood and reconstructing the live set.
+}
+\details{
+The two runs must use the same prior variable names and LRPS method.
+If \code{first_update} or \code{update_interval} differ between runs, a warning is
+thrown and defaults are re-computed from the merged \code{nlive}.
+
+The merged run contains all dead points from both runs, ordered by
+log-likelihood, and the live set is reconstructed from the remaining points.
+This results in a single nested sampling run with \code{nlive} equal to the sum of
+the two original runs.
+
+Note that \code{niter} is not simply the sum of the two original runs' \code{niter}
+values; rather, the merged run's \code{niter} is equal to the total number of
+samples drawn before either run's termination condition was met. This means
+that the merged run's \code{niter} reflects the number of iterations it would take
+to reach the same likelihood contour as the least advanced of the two
+original runs.
+
+If the two runs do not share the same RNG seed, the merged result stores
+\code{NA} as the seed.
+}
+\examples{
+prior <- create_uniform_prior(lower = c(-1, -1), upper = 1)
+ll_fn <- function(x) -sum(x^2)
+
+run1 <- generate(
+  ernest_sampler(ll_fn, prior, nlive = 100, seed = 24),
+  max_iterations = 100
+)
+run2 <- generate(
+  ernest_sampler(ll_fn, prior, nlive = 300, seed = 42),
+  max_iterations = 100
+)
+merged <- merge(run1, run2)
+summary(merged)
+}
