+++ NAMESPACE
@@ -18,6 +18,7 @@ S3method(format,unif_cube)
 S3method(format,unif_ellipsoid)
 S3method(generate,ernest_run)
 S3method(generate,ernest_sampler)
+S3method(merge,ernest_run)
 S3method(new_ernest_run,ernest_run)
 S3method(new_ernest_run,ernest_sampler)
 S3method(plot,ernest_estimate)
