### 2026-03-20 4fa1769
New merge method.
