+++ tests/testthat/test-merge.R
@@ -1,0 +1,125 @@
+describe("butcher", {
+  it("catches errors")
+
+  it("extracts the correct elements from an ernest_run object", {
+    b <- butcher(example_run, first_id = 1)
+    expect_named(
+      b,
+      c("log_lik", "id", "evaluations", "birth_lik", "unit")
+    )
+    expect_mapequal(
+      b[c("log_lik", "id", "evaluations", "birth_lik")],
+      example_run$weights[c("log_lik", "id", "evaluations", "birth_lik")]
+    )
+  })
+
+  test_that("remaps ids", {
+    first_id <- 51L
+    b <- butcher(example_run, first_id = first_id)
+    expect_identical(sort(unique(b$id)), c(51:1050))
+
+    gapped_ids <- example_run
+    example_run$weights$id[which(
+      example_run$weights$id > 200
+    )] <- example_run$weights$id[which(example_run$weights$id > 200)] + 1000L
+    b <- butcher(gapped_ids, first_id = 1)
+    expect_identical(sort(unique(b$id)), c(1L:1000L))
+  })
+})
+
+describe("merge_sampler", {
+  x <- ernest_sampler(
+    log_lik = gaussian_blobs$log_lik,
+    prior = gaussian_blobs$prior
+  )
+
+  it("catches errors", {
+    expect_error(
+      merge_sampler(x, list()),
+      "must be an object with class ernest_sampler"
+    )
+    expect_error(
+      merge_sampler(1, x),
+      "must be an object with class ernest_sampler"
+    )
+    sampler_3d <- ernest_sampler(
+      log_lik = gaussian_3D$log_lik,
+      prior = gaussian_3D$prior
+    )
+    expect_error(
+      merge_sampler(x, sampler_3d),
+      "must have the same prior variable names"
+    )
+    sampler_alt <- ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
+      prior = gaussian_blobs$prior,
+      sampler = unif_cube()
+    )
+    expect_error(
+      merge_sampler(x, sampler_alt),
+      "must have the same LRPS method"
+    )
+  })
+
+  it("combines compatible samplers", {
+    y <- ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
+      prior = gaussian_blobs$prior
+    )
+
+    z <- merge_sampler(x, y)
+    expect_identical(z$nlive, 1000L)
+    expect_identical(z$first_update, 1250L)
+    expect_identical(z$update_interval, 750L)
+  })
+
+  it("warns and uses default update values when they differ", {
+    x <- ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
+      prior = gaussian_blobs$prior,
+      first_update = 100,
+      update_interval = 50,
+      seed = 24
+    )
+    y <- ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
+      prior = gaussian_blobs$prior,
+      first_update = 150,
+      update_interval = 75,
+      seed = 42
+    )
+
+    expect_snapshot(z <- merge_sampler(x, y))
+    expect_identical(z$first_update, 2500L)
+    expect_identical(z$update_interval, 1500L)
+    expect_identical(attr(z, "seed"), NA_integer_)
+  })
+})
+
+test_that("merging two runs", {
+  run1 <- generate(
+    ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
+      prior = gaussian_blobs$prior,
+      seed = 24
+    ),
+    max_iterations = 100
+  )
+  run2 <- generate(
+    ernest_sampler(
+      log_lik = gaussian_blobs$log_lik,
[…]
