### 2026-03-20 67c22f0
Reworked control function for generate/nested_sampling_impl
