### 2026-03-20 e1bef72
New info method for nested_sampling_impl.
