### 2026-03-20 b555fcc
New info method for nested_sampling_impl.
