+++ R/generate.R
@@ -104,15 +104,22 @@ generate.ernest_sampler <- function(
     show_progress <- getOption("rlib_message_verbosity", "default") != "quiet"
   }
   check_bool(show_progress)
-  control <- set_run_control(
+  control <- get_run_control(
     x,
     max_iterations,
     max_evaluations,
     min_logz
   )
   x <- compile(x, ...)
+  info <- get_sampler_info(x)
 
-  results <- nested_sampling_impl(x, control, show_progress = show_progress)
+  results <- nested_sampling_impl(
+    live_env = x$run_env,
+    lrps = x$lrps,
+    sampler_info = info,
+    control = control,
+    show_progress = show_progress
+  )
   new_ernest_run(x, results)
 }
 
@@ -144,14 +151,21 @@ generate.ernest_run <- function(
     ))
   }
 
-  control <- set_run_control(
+  control <- get_run_control(
     x,
     max_iterations,
     max_evaluations,
     min_logz
   )
+  info <- get_sampler_info(x)
 
-  results <- nested_sampling_impl(x, control, show_progress = show_progress)
+  results <- nested_sampling_impl(
+    live_env = x$run_env,
+    lrps = x$lrps,
+    sampler_info = info,
+    control = control,
+    show_progress = show_progress
+  )
   new_ernest_run(x, results)
 }
 
@@ -167,7 +181,7 @@ generate.ernest_run <- function(
 #' @return A named list containing `max_iterations`, `max_evaluations`,
 #' `min_logz`, `last_criterion`, `log_z`, `log_vol`, `cur_iter`, and `cur_eval`.
 #' @noRd
-set_run_control <- function(
+get_run_control <- function(
   x,
   max_iterations,
   max_evaluations,
@@ -258,5 +272,20 @@ set_run_control <- function(
     log_z = log_z,
     cur_iter = as.integer(cur_iter),
     cur_eval = as.integer(cur_eval)
+  )
+}
+
+#' Extract sampler information for use in nested sampling.
+#'
+#' @param x An `ernest_sampler` or `ernest_run` object.
+#' @return A named list containing `seed`, `nlive`, `first_update`,
+#' and `update_interval`.
+#' @noRd
+get_sampler_info <- function(x) {
+  list(
+    seed = attr(x, "seed"),
+    nlive = x$nlive,
+    first_update = x$first_update,
+    update_interval = x$update_interval
   )
 }
