### 2026-03-23 8a0d41d
Early parallel routines for generate.
