+++ R/parallel.R
@@ -1,0 +1,176 @@
+#' @param x The ernest_sampler or ernest_run object.
+#' @param sampler_info A list containing information about the sampler.
+#' @param control parameters for the nested sampling run, generated from
+#' `set_run_control()`.
+#' @param show_progress Logical. If `TRUE`, displays a progress bar during
+#' sampling.
+#'
+#' @noRd
+p_generate <- function(
+  x,
+  sampler_info,
+  run_control,
+  show_progress,
+  allow_par,
+  call = caller_env()
+) {
+  check_installed("mirai", "allow_par", call = call)
+  mirai::require_daemons(call = call)
+  nlive <- x$nlive
+  hint <- NULL
+  workers <- if (isTRUE(allow_par)) {
+    hint <- "Have you changed the number of daemons set by {.pkg mirai}?"
+    rep(na_int, mirai::info()[["connections"]])
+  } else {
+    hint <- "Have you changed how the run is parallelized with `allow_par`?"
+    vctrs::vec_cast(allow_par, integer(), call = call)
+  }
+  nlive_threaded <- thread_nlive(nlive, workers, call)
+  thread_envs <- split_live(x$run_env, nlive_threaded, sampler_info, call)
+
+  m <- mirai::mirai_map(
+    thread_envs$live,
+    \(thread) {
+      fn(
+        thread$env,
+        lrps = l,
+        thread$info,
+        control = rc,
+        show_progress = FALSE
+      )
+    },
+    fn = nested_sampling_impl,
+    l = x$lrps,
+    rc = run_control
+  )
+
+  results <- mirai::collect_mirai(
+    m,
+    options = c(".stop", if (show_progress) ".progress" else NULL)
+  )
+  result <- collect_results(results, thread_envs$split)
+  collect_live(x$run_env, thread_envs$live, thread_envs$split)
+  new_ernest_run(x, result)
+}
+
+#' Get the nlive for each worker.
+#'
+#' @param nlive The total nlive of the sampler.
+#' @param workers An integer vector containing nlive per worker or NA.
+#' @param hint A string printed when an error is thrown.
+#' @param call Calling environment for errors.
+#'
+#' @returns A validated integer vector, whose length is the number of
+#' runs and each element is that run's nlive.
+#' @noRd
+thread_nlive <- function(nlive, workers, hint, call = caller_env()) {
+  if (vctrs::vec_is_empty(workers)) {
+    cli::cli_abort(
+      c("At least one worker must be specified.", "i" = hint),
+      call = call
+    )
+  }
+  workers <- if (all(is.na(workers))) {
+    rep(nlive %/% length(workers), length(workers))
+  } else if (any(!is.finite(workers)) || any(workers < 1L)) {
+    cli::cli_abort(
+      c("parallel runs must each contain at least one live point", "i" = hint),
+      call = call
+    )
+  } else {
+    workers
+  }
+  if (sum(workers) > nlive) {
+    cli::cli_abort(
+      c(
+        "parallel runs must contain a total of {nlive} live points",
+        "i" = hint
+      ),
+      call = call
+    )
+  }
+  workers[[1]] <- workers[[1]] + (nlive - sum(workers))
+  as.integer(workers)
+}
+
+#' Split the live set into a series of threads.
+#'
+#' @param live_env The original live set, already validated.
+#' @param workers Integer vector describing how points are to be
+#' allocated
+#' @param info Sampler info, which is broken up into per-thread control lists.
+#' @param call Error info.
+#'
[…]
