### 2026-03-23 4d94445
`merge` matches nested_sampling_impl output.
