+++ R/merge.R
@@ -7,7 +7,7 @@
 #' specifications.
 #' @inheritParams rlang::args_dots_empty
 #'
-#' @returns [[ernest_run]] containing merged dead points and live points.
+#' @returns [[ernest_run]] A run containing the merged results from `x` and `y`.
 #'
 #' @details
 #' The two runs must use the same prior variable names and LRPS method.
@@ -48,45 +48,75 @@ merge.ernest_run <- function(x, y, ...) {
   z <- merge_sampler(x, y)
   # Merge results
   list_x <- butcher(x)
-  list_y <- butcher(y, first_id = max(list_x$id) + 1)
-  lik_unordered <- c(list_x$log_lik, list_y$log_lik)
-  lik_ordered <- order(lik_unordered)
-  xy_ordered <- data.frame(
-    log_lik = lik_unordered[lik_ordered],
-    id = c(list_x$id, list_y$id)[lik_ordered],
-    evaluations = c(list_x$evaluations, list_x$evaluations)[lik_ordered],
-    birth_lik = c(list_x$birth_lik, list_y$birth_lik)[lik_ordered],
-    .iter = seq_along(c(list_x$birth_lik, list_y$birth_lik))
+  list_y <- butcher(y, first_id = max(list_c(list_x$dead_id)) + 1L)
+  results <- merge_results(list_x, list_y)
+  # Reform sampler
+  env_bind(z$run_env, !!!results$live)
+  z <- refresh_ernest_sampler(z)
+  new_ernest_run(z, results$dead)
+}
+
+#' Merge the dead points from multiple nested sampling runs.
+#'
+#' @param ... Lists of nested sampling results, each containing "unit",
+#' "log_lik", "id", "evals", and "birth_lik" elements.
+#' @param .call Information about the calling environment for error messages.
+#' @returns A list containing merged "live" and "dead" points from the input
+#' runs.
+#'
+#' @noRd
+merge_results <- function(..., .call = caller_env()) {
+  runs <- list2(...)
+  flattened <- lapply(runs, \(r) parse_results(r))
+  results <- list(
+    "unit" = do.call(rbind, lapply(flattened, `[[`, "unit")),
+    "log_lik" = do.call(c, lapply(flattened, `[[`, "log_lik")),
+    "id" = do.call(c, lapply(flattened, `[[`, "id")),
+    "evals" = do.call(c, lapply(flattened, `[[`, "evals")),
+    "birth_lik" = do.call(c, lapply(flattened, `[[`, "birth_lik"))
   )
-  xy_unit <- rbind(list_x$unit, list_y$unit)[lik_ordered, , drop = FALSE]
-  # Get dead points
-  last_death <- match(TRUE, xy_ordered$evaluations == 0) - 1
-  max_lik <- xy_ordered$log_lik[[last_death]]
-  results <- {
-    xy_dead <- xy_ordered[seq(last_death), ]
-    list(
-      "dead_unit" = xy_unit[seq(last_death), ],
-      "dead_log_lik" = xy_dead$log_lik,
-      "dead_id" = xy_dead$id,
-      "dead_evals" = xy_dead$evaluations,
-      "dead_birth" = xy_dead$birth_lik
-    )
-  }
+  order <- order(results$log_lik)
+  results <- list(
+    "unit" = results$unit[order, , drop = FALSE],
+    "log_lik" = results$log_lik[order],
+    "id" = results$id[order],
+    "evals" = results$evals[order],
+    "birth_lik" = results$birth_lik[order]
+  )
   # Get live points
+  first_live <- match(TRUE, results$evals == 0L)
+  min_live <- results$log_lik[[first_live]]
+  ordered <- data.frame(
+    log_lik = results$log_lik,
+    id = results$id,
[…]
