### 2026-03-23 3e47fd1
Parallel generate now takes advantage of merge.
