### 2026-03-24 3b1f3f4
Rename to 'live_env' for clarity; repaired issue with log-evidence SE calculation.
