### 2026-03-24 71a27da
Simpler run-class constructors to allow for easier implementation of merge and parallel.
