+++ R/merge.R
@@ -45,76 +45,65 @@
 #' summary(merged)
 #' @export
 merge.ernest_run <- function(x, y, ...) {
+  check_class(y, "ernest_run")
   z <- merge_sampler(x, y)
   # Merge results
-  list_x <- butcher(x)
-  list_y <- butcher(y, first_id = max(list_c(list_x$dead_id)) + 1L)
-  results <- merge_results(list_x, list_y)
+  list_x <- butcher_run(x, keep_live = TRUE)
+  list_y <- butcher_run(y, keep_live = TRUE)
+  lists_xy <- merge_ids(list_x, list_y)
+  results <- merge_results(lists_xy)
   # Reform sampler
   env_bind(z$live_env, !!!results$live)
   z <- refresh_ernest_sampler(z)
   new_ernest_run(z, results$dead)
 }
 
-#' Merge the dead points from multiple nested sampling runs.
-#'
-#' @param ... Lists of nested sampling results, each containing "unit",
-#' "log_lik", "id", "evals", and "birth_lik" elements.
-#' @param .call Information about the calling environment for error messages.
-#' @returns A list containing merged "live" and "dead" points from the input
-#' runs.
+#' Correct the IDs of two merged runs.
 #'
 #' @noRd
-merge_results <- function(..., .call = caller_env()) {
+merge_ids <- function(...) {
   runs <- list2(...)
-  flattened <- lapply(runs, \(r) parse_results(r))
-  results <- list(
-    "unit" = do.call(rbind, lapply(flattened, `[[`, "unit")),
-    "log_lik" = do.call(c, lapply(flattened, `[[`, "log_lik")),
-    "id" = do.call(c, lapply(flattened, `[[`, "id")),
-    "evals" = do.call(c, lapply(flattened, `[[`, "evals")),
-    "birth_lik" = do.call(c, lapply(flattened, `[[`, "birth_lik"))
+  # Correct the IDS
+  nlives <- vapply(runs, \(x) length(unique(x$id)), integer(1))
+  .mapply(
+    \(x, start) {
+      x$id <- vctrs::vec_group_id(x$id) + start
+      x
+    },
+    dots = list(runs, c(0, cumsum(nlives)[-length(nlives)])),
+    MoreArgs = NULL
   )
-  order <- order(results$log_lik)
-  results <- list(
-    "unit" = results$unit[order, , drop = FALSE],
-    "log_lik" = results$log_lik[order],
-    "id" = results$id[order],
-    "evals" = results$evals[order],
-    "birth_lik" = results$birth_lik[order]
-  )
+}
+
+#' Merge the dead points from multiple nested sampling runs.
+#'
+#' @param runs A list of nested sampling results from bucher_run. These
+#' should have unique IDs.
+#'
+#' @returns Two lists of "dead" and "live" run information.
+#' @noRd
+merge_results <- function(runs) {
+  all_runs <- vctrs::vec_rbind(!!!lapply(runs, \(x) vctrs::data_frame(!!!x)))
+  all_runs <- all_runs[order(all_runs$log_lik), ]
   # Get live points
-  first_live <- match(TRUE, results$evals == 0L)
-  min_live <- results$log_lik[[first_live]]
-  ordered <- data.frame(
-    log_lik = results$log_lik,
-    id = results$id,
-    .iter = seq_along(results$log_lik)
-  )
-  live_idx <- vapply(
-    split(ordered, ordered$id),
-    \(df) df$.iter[match(TRUE, df$log_lik >= min_live)],
-    integer(1)
-  )
-  live <- list(
-    unit = results$unit[live_idx, , drop = FALSE],
[…]
