### 2026-03-25 307bb88
Parallel now correctly drops live points.
