### 2026-03-25 19caa7d
New ernest_rcrd class for easier ernest_run merging and construction.
