+++ R/run-class.R
@@ -13,22 +13,14 @@ new_ernest_run <- function(x, results) {
 #' @export
 #' @noRd
 new_ernest_run.ernest_sampler <- function(x, results) {
-  parsed <- parse_results(results, x$live_env)
-  new_ernest_run_(x, parsed)
+  new_ernest_run_(x, results)
 }
 
 #' @export
 #' @noRd
 new_ernest_run.ernest_run <- function(x, results) {
-  prev_run <- butcher_run(x, keep_live = FALSE)
-  parsed <- parse_results(results, x$live_env)
-
-  parsed$dead <- vctrs::vec_rbind(
-    vctrs::data_frame(!!!prev_run),
-    vctrs::data_frame(!!!parsed$dead)
-  )
-  parsed$niter <- x$niter + parsed$niter
-  new_ernest_run_(x, parsed)
+  prev_run <- as_ernest_rcrd(x, keep_live = FALSE)
+  new_ernest_run_(x, vctrs::vec_c(prev_run, results))
 }
 
 #' Form the new_ernest_run from samples from the current and previous runs
@@ -42,36 +34,36 @@ new_ernest_run.ernest_run <- function(x, results) {
 #' @return The object described by generate.
 #' @noRd
 new_ernest_run_ <- function(x, parsed) {
-  all_samples <- bind_dead_live(parsed, x$nlive)
-  samples_unit <- all_samples$unit
-  colnames(samples_unit) <- x$prior$names
-  samples <- t(apply(samples_unit, 1, x$prior$fn)) # TODO: Try vectorized?
-  colnames(samples) <- x$prior$names
-
+  all_samples <- c(parsed, extract_live_points(x$live_env))
+  niter <- sum(field(all_samples, "evals") != 0L)
   integration <- compute_integral(
-    all_samples$log_lik,
-    get_log_vol(x$nlive, niter = parsed$niter)
+    field(all_samples, "log_lik"),
+    get_log_vol(x$nlive, niter = niter)
   )
+  unit <- as.list(all_samples)[["unit"]]
+  colnames(unit) <- x$prior$names
+  original <- t(apply(unit, 1, x$prior$fn))
+  colnames(original) <- x$prior$names
 
-  result_elem <- list2(
-    "niter" = parsed$niter,
-    "neval" = sum(all_samples$evals),
+  result_elem <- list(
+    "niter" = niter,
+    "neval" = sum(field(all_samples, "evals")),
     "log_evidence" = tail(integration$log_evidence, 1L),
     "log_evidence_err" = sqrt(tail(integration$log_evidence_var, 1L)),
     "information" = tail(integration$information, 1L),
     "samples" = list(
-      "original" = samples,
-      "unit_cube" = samples_unit
+      "original" = original,
+      "unit_cube" = unit
     ),
     "weights" = vctrs::df_list(
-      "id" = as.integer(all_samples$id),
-      "evaluations" = as.integer(all_samples$evals),
-      "log_lik" = all_samples$log_lik,
+      "id" = field(all_samples, "id"),
+      "evaluations" = as.integer(field(all_samples, "evals")),
+      "log_lik" = integration$log_lik,
       "log_weight" = integration$log_w,
       "imp_weight" = exp(
         integration$log_w - tail(integration$log_evidence, 1L)
       ),
-      "birth_lik" = all_samples$birth_lik
+      "birth_lik" = field(all_samples, "birth_lik")
     )
   )
 
@@ -116,103 +108,6 @@ print.ernest_run <- function(x, ...) {
     "* Information: {h}"
   ))
   invisible(x)
-}
-
-# HELPERS FOR ERNEST_RUN-----
-
-#' Parse the results from nested_sampling_impl into a list
-#'
[…]
