### 2026-03-25 0f8469f
Fixed off-by-one error in slice_rectangle; fixed vec_ptype2 for ernest_rcrd.
