### 2026-03-26 cc4512c
New stopping criterion validation for generate.
