+++ R/generate.R
@@ -108,8 +108,7 @@ generate.ernest_sampler <- function(
     show_progress <- getOption("rlib_message_verbosity", "default") != "quiet"
   }
   check_bool(show_progress)
-  control <- get_run_control(
-    x,
+  control <- new_generate_control(
     max_iterations,
     max_evaluations,
     min_logz
@@ -157,20 +156,15 @@ generate.ernest_run <- function(
   check_bool(show_progress)
   x <- compile(x, ...)
   if (inherits_only(x, "ernest_sampler")) {
-    return(generate(
-      x,
-      max_iterations = max_iterations,
-      max_evaluations = max_evaluations,
-      min_logz = min_logz,
-      show_progress = show_progress
-    ))
+    return(NextMethod())
   }
 
-  control <- get_run_control(
-    x,
+  x_rcrd <- as_ernest_rcrd(x)
+  control <- new_generate_control(
     max_iterations,
     max_evaluations,
-    min_logz
+    min_logz,
+    prev_run = x_rcrd
   )
   info <- get_sampler_info(x)
 
@@ -190,29 +184,32 @@ generate.ernest_run <- function(
 
 #' Generate and validate stopping criteria.
 #'
-#' @param x Object used for nested sampling.
 #' @param max_iterations Maximum number of iterations to perform.
 #' @param max_evaluations Maximum number of likelihood function evals.
 #' @param min_logz Minimum log-ratio between current estimated evidence and
 #' remaining evidence.
+#' @param prev_run The points from a previous run; NULL if no points yet
+#' exist.
 #' @param call Environment for error reporting.
 #'
 #' @return A named list containing `max_iterations`, `max_evaluations`,
 #' `min_logz`, `last_criterion`, `log_z`, `log_vol`, `cur_iter`, and `cur_eval`.
 #' @noRd
-get_run_control <- function(
-  x,
+new_generate_control <- function(
   max_iterations,
   max_evaluations,
   min_logz,
+  prev_run = NULL,
   call = caller_env()
 ) {
   check_number_whole(max_iterations, min = 1, allow_null = TRUE, call = call)
   check_number_whole(max_evaluations, min = 1, allow_null = TRUE, call = call)
   check_number_decimal(min_logz, min = 0, call = call)
-  no_stopping <- identical(min_logz, 0) &&
-    is.null(max_iterations) &&
+  no_stopping <- all(
+    identical(min_logz, 0),
+    is.null(max_iterations),
     is.null(max_evaluations)
+  )
   if (no_stopping) {
     cli::cli_abort(
       c(
@@ -226,7 +223,7 @@ get_run_control <- function(
   max_iterations <- max_iterations %||% .Machine$integer.max
   max_evaluations <- max_evaluations %||% .Machine$integer.max
 
-  if (inherits_only(x, "ernest_sampler")) {
+  if (is.null(prev_run)) {
     return(list(
       max_iterations = as.integer(max_iterations),
       max_evaluations = as.integer(max_evaluations),
@@ -239,32 +236,34 @@ get_run_control <- function(
     ))
   }
 
-  cur_iter <- x$niter
-  cur_eval <- x$neval
+  nlive <- vctrs::vec_unique_count(field(prev_run, "id"))
+  niter <- match(0L, field(prev_run, "evals")) - 1L
+  neval <- sum(field(prev_run, "evals"))
   prev_integration <- compute_integral(
-    x$weights$log_lik,
-    get_log_vol(x$nlive, niter = cur_iter)
+    field(prev_run, "log_lik"),
+    get_log_vol(nlive, niter = niter)
   )
-  last_criterion <- prev_integration$log_lik[cur_iter]
-  log_z <- prev_integration$log_evidence[cur_iter]
[…]
