### 2026-03-26 9161fe5
New compile workflow using vctrs methods.
