+++ tests/testthat/test-compile.R
@@ -1,19 +1,14 @@
 withr::local_seed(42)
 
-sampler <- NULL
-test_that("Set up sampler", {
-  expect_no_error(
-    sampler <<- ernest_sampler(
-      log_lik = gaussian_blobs$log_lik,
-      prior = gaussian_blobs$prior,
-      seed = 42
-    )
+describe("new_live_set", {
+  sampler <- ernest_sampler(
+    log_lik = gaussian_blobs$log_lik,
+    prior = gaussian_blobs$prior,
+    seed = 42
   )
-})
 
-describe("create_live", {
   it("generates the live set correctly", {
-    result <- create_live(sampler$lrps, 10)
+    result <- new_live_set(sampler$lrps, 10)
     expect_equal(dim(result$unit), c(10, 2))
     expect_equal(
       apply(
@@ -24,183 +19,166 @@ describe("create_live", {
       result$log_lik
     )
   })
-  it("gives informative error when unit_log_fn fails completely", {
+
+  it("gives informative errors when unit_log_fn fails", {
     bad_lik <- new_rwmh_cube(
       unit_log_fn = \(x) stop("Bad Likelihood!"),
       n_dim = 2L
     )
-    expect_error(create_live(bad_lik, 10), "Bad Likelihood!")
+    expect_error(new_live_set(bad_lik, 10), "Bad Likelihood!")
   })
 })
 
-#' @srrstats {BS2.1a} check_set_live() ensures that log_lik produces an expected
-#' output given some output from the prior transformation, ensuring the two
-#' are commensurate. It is called by compile and by ernest_sampler.
-describe("check_live_set", {
-  reset_live <- function() {
-    env_bind(
-      sampler$live_env,
-      unit = matrix(runif(1000), nrow = 500, ncol = 2),
-      log_lik = seq(-10, -1, length.out = 500),
-      birth_lik = rep(-Inf, 500)
-    )
+#' @srrstats {BS2.1a} write_live_set() validates live set components and ensures
+#' that log_lik produces expected output given prior transformation, ensuring
+#' commensurate quantities. It is called by compile and by ernest_sampler.
+describe("write_live_set", {
+  sampler <- ernest_sampler(
+    log_lik = gaussian_blobs$log_lik,
+    prior = gaussian_blobs$prior,
+    seed = 42,
+    nlive = 500
+  )
+
+  make_live <- function(
+    unit = matrix(runif(1000), nrow = 500, ncol = 2),
+    log_lik = seq(-10, -1, length.out = 500),
+    birth_lik = rep(-Inf, 500)
+  ) {
+    list(unit = unit, log_lik = log_lik, birth_lik = birth_lik)
   }
 
-  it("passes with valid live set", {
-    reset_live()
-    expect_silent(check_live_set(sampler))
+  it("accepts valid live set and binds to live_env", {
+    live <- make_live()
+    result <- write_live_set(live, sampler)
+    expect_true(exists("unit", envir = result))
+    expect_true(exists("log_lik", envir = result))
+    expect_true(exists("birth_lik", envir = result))
+    expect_equal(nrow(result$unit), 500)
+    expect_equal(length(result$log_lik), 500)
+    expect_equal(length(result$birth_lik), 500)
   })
 
-  it("errors if unit is not a matrix", {
-    reset_live()
-    env_bind(sampler$live_env, unit = runif(20))
+  it("errors if unit doesn't match n_dim", {
+    live <- make_live(unit = matrix(runif(1500), nrow = 500, ncol = 3))
     expect_error(
-      check_live_set(sampler),
-      "`unit` must be a matrix, not a double vector."
-    )
-
-    reset_live()
-    env_bind(
[…]
