+++ R/utils.R
@@ -54,86 +54,6 @@ check_class <- function(
   )
 }
 
-#' Check that an object is a double matrix
-#'
-#' Validates that an object is a double matrix with specified dimensions
-#' and within given bounds.
-#'
-#' @param x An object to check.
-#' @param nrow Expected number of rows.
-#' @param ncol Expected number of columns.
-#' @param lower Numeric. Exclusive lower bounds, recycled to length `ncol`.
-#' @param upper Numeric. Exclusive upper bounds, recycled to length `ncol`.
-#' @param finite Logical. If TRUE, checks that all values are finite.
-#' @param ... Additional arguments passed to error handlers.
-#' @param arg Argument name for error messages.
-#' @param call Call environment for error messages.
-#'
-#' @return Returns NULL invisibly if `x` is a valid double matrix,
-#' otherwise throws an informative error.
-#' @noRd
-check_matrix <- function(
-  x,
-  nrow,
-  ncol,
-  lower = -Inf,
-  upper = Inf,
-  finite = TRUE,
-  ...,
-  arg = caller_arg(x),
-  call = caller_env()
-) {
-  if (!is.matrix(x)) {
-    stop_input_type(x, "a matrix", arg = arg, call = call)
-  }
-
-  if (typeof(x) != "double") {
-    stop_input_type(
-      x,
-      "a double matrix",
-      arg = arg,
-      call = call
-    )
-  }
-
-  if (nrow(x) != nrow) {
-    cli::cli_abort("`{arg}` must have {nrow} rows, not {nrow(x)}.", call = call)
-  }
-  if (ncol(x) != ncol) {
-    cli::cli_abort(
-      "`{arg}` must have {ncol} columns, not {ncol(x)}.",
-      call = call
-    )
-  }
-
-  if (finite && any(!is.finite(x))) {
-    cli::cli_abort(
-      "`{arg}` must contain no nonfinite values.",
-      call = call
-    )
-  }
-
-  if (identical(lower, -Inf) && identical(upper, Inf)) {
-    return(invisible(NULL))
-  }
-
-  bounds <- vctrs::vec_recycle_common(
-    lower = lower,
-    upper = upper,
-    .size = ncol
-  )
-  mins <- matrixStats::colMins(x)
-  maxes <- matrixStats::colMaxs(x)
-
-  if (any(bounds$lower > mins)) {
-    cli::cli_abort("`{arg}` must respect the lower bounds.")
-  }
-  if (any(bounds$upper < maxes)) {
-    cli::cli_abort("`{arg}` must respect the upper bounds.")
-  }
-  invisible(NULL)
-}
-
 #' Check that a list has unique, non-empty names
 #'
 #' Validates that all elements of a list are named and that names are unique.
@@ -172,18 +92,6 @@ check_unique_names <- function(
   }
 
   invisible(NULL)
-}
-
-#' Inject a list into a vector
-#'
-#' Concatenates the elements of a list into a single vector.
-#'
-#' @param x A list to concatenate.
-#'
-#' @return A vector containing all elements of the list.
-#' @noRd
-list_c <- function(x) {
-  inject(c(!!!x))
 }
 
 # Helpers for computing and reporting results -----
