### 2026-03-26 7b06116
Test coverage, deleted unused cpp export.
