+++ tests/testthat/test-visualize.R
@@ -18,10 +18,8 @@ test_that("visualize validates parameters", {
   )
 })
 
+skip_plot_snapshot()
 test_that("Density plot", {
-  skip_extended()
-  skip_on_cran()
-  skip_on_ci()
   data(example_run)
   vdiffr::expect_doppelganger(
     "visualize-basic",
@@ -45,9 +43,6 @@ test_that("Density plot", {
 })
 
 test_that("Trace plot", {
-  skip_extended()
-  skip_on_cran()
-  skip_on_ci()
   data(example_run)
   vdiffr::expect_doppelganger(
     "trace-basic",
