### 2026-03-26 dc32653
Man and example checks done.
