+++ NAMESPACE
@@ -73,7 +73,6 @@ export(unif_ellipsoid)
 export(update_lrps)
 export(visualize)
 import(rlang)
-importFrom(cli,cli_warn)
 importFrom(cli,col_green)
 importFrom(cli,pb_current)
 importFrom(cli,pb_elapsed)
