### 2026-03-27 891c96e
extract_live_set now sorts the set properly; birth_lik values are now correctly written.
