### 2026-03-27 4304198
Reworked parallel and merge with correct indexing.
