+++ tests/testthat/test-merge.R
@@ -1,96 +1,99 @@
-describe("merge_sampler", {
-  x <- ernest_sampler(
+run1 <- generate(
+  ernest_sampler(
     log_lik = gaussian_blobs$log_lik,
-    prior = gaussian_blobs$prior
+    prior = gaussian_blobs$prior,
+    seed = 24
+  ),
+  max_iterations = 100
+)
+run2 <- generate(
+  ernest_sampler(
+    log_lik = gaussian_blobs$log_lik,
+    prior = gaussian_blobs$prior,
+    nlive = 300,
+    seed = 42
+  ),
+  max_iterations = 100
+)
+
+test_that("reindex_runs remaps IDs and preserves record fields", {
+  x1 <- as_ernest_rcrd(run1)
+  x2 <- as_ernest_rcrd(run2)
+  out <- reindex_runs(x1, x2)
+
+  n1 <- length(x1)
+  n2 <- length(x2)
+  nlive1 <- vctrs::vec_unique_count(vctrs::field(x1, "id"))
+  nlive2 <- vctrs::vec_unique_count(vctrs::field(x2, "id"))
+
+  expect_length(out, 2)
+  expect_identical(sort(unique(vctrs::field(out[[1]], "id"))), seq(nlive1))
+  expect_identical(
+    sort(unique(vctrs::field(out[[2]], "id"))),
+    seq(nlive1 + 1, nlive1 + nlive2)
   )
 
-  it("catches errors", {
-    expect_error(
-      merge_sampler(x, list()),
-      "must be an object with class ernest_sampler"
-    )
-    expect_error(
-      merge_sampler(1, x),
-      "must be an object with class ernest_sampler"
-    )
-    sampler_3d <- ernest_sampler(
-      log_lik = gaussian_3D$log_lik,
-      prior = gaussian_3D$prior
-    )
-    expect_error(
-      merge_sampler(x, sampler_3d),
-      "must have the same prior variable names"
-    )
-    sampler_alt <- ernest_sampler(
-      log_lik = gaussian_blobs$log_lik,
-      prior = gaussian_blobs$prior,
-      sampler = unif_cube()
-    )
-    expect_error(
-      merge_sampler(x, sampler_alt),
-      "must have the same LRPS method"
-    )
-  })
+  expect_identical(
+    vctrs::field(out[[1]], "log_lik"),
+    vctrs::field(x1, "log_lik")
+  )
+  expect_identical(
+    vctrs::field(out[[2]], "log_lik"),
+    vctrs::field(x2, "log_lik")
+  )
+})
 
-  it("combines compatible samplers", {
-    y <- ernest_sampler(
-      log_lik = gaussian_blobs$log_lik,
-      prior = gaussian_blobs$prior
-    )
+test_that("merge_results errors when IDs are duplicated across runs", {
+  x1 <- as_ernest_rcrd(run1)
+  x2 <- as_ernest_rcrd(run2)
 
-    z <- merge_sampler(x, y)
-    expect_identical(z$nlive, 1000L)
-    expect_identical(z$first_update, 1250L)
-    expect_identical(z$update_interval, 750L)
-  })
+  expect_error(
+    merge_results(x1, x2),
+    "must contain unique IDs"
+  )
+})
 
-  it("uses default update values when they differ", {
-    x <- ernest_sampler(
-      log_lik = gaussian_blobs$log_lik,
-      prior = gaussian_blobs$prior,
-      first_update = 100,
-      update_interval = 50,
-      seed = 24
-    )
-    y <- ernest_sampler(
-      log_lik = gaussian_blobs$log_lik,
-      prior = gaussian_blobs$prior,
-      first_update = 150,
[…]
