+++ R/generate.R
@@ -120,10 +120,9 @@ generate.ernest_sampler <- function(
   if (!isFALSE(allow_par)) {
     p_generate(
       x,
-      sampler_info = info,
-      run_control = control,
-      show_progress = show_progress,
-      allow_par = allow_par
+      parent_info = info,
+      parent_control = control,
+      show_progress = show_progress
     )
   } else {
     results <- nested_sampling_impl(
@@ -170,7 +169,7 @@ generate.ernest_run <- function(
   info <- get_sampler_info(x)
 
   if (!isFALSE(allow_par)) {
-    p_generate(x, info, control, show_progress, allow_par)
+    p_generate(x, parent_info = info, parent_control = control, show_progress)
   } else {
     results <- nested_sampling_impl(
       live_env = x$live_env,
