### 2026-03-27 2fc069a
Simple bootstrapping method added.
