### 2026-03-30 e09bcd2
Faster learn subroutine
