+++ tests/testthat/test-calculate.R
@@ -5,7 +5,16 @@ withr::local_seed(42)
 gold <- readRDS(test_path("calculate-gold.rds"))
 
 test_that("Helpers produce as expected", {
-  expect_equal(drop(get_log_vol(250, 2750)), gold$log_volume)
+  expect_equal(
+    get_points(c(10, 9, 8, 7, 6, 5, 4), 3),
+    c(3, 3, 3, 3, 3, 2, 1)
+  )
+  expect_equal(
+    get_points(c(10, 10, 9, 8, 7, 7, 7), 3),
+    c(3, 2, 3, 3, 3, 2, 1)
+  )
+
+  expect_equal(drop(get_log_vol(gold$log_lik, nlive = 250)), gold$log_volume)
   calc <- get_log_w(gold$log_lik, gold$log_volume)
   expect_equal(drop(calc$log_weight), gold$log_weight)
   expect_equal(drop(calc$log_evidence), gold$log_evidence)
@@ -13,7 +22,7 @@ test_that("Helpers produce as expected", {
 
 test_that("Simulated log vols do not diverge from mean estimates", {
   set.seed(42)
-  log_vol <- get_log_vol(250, 2750, ndraws = 4000)
+  log_vol <- get_log_vol(gold$log_lik, 250, ndraws = 4000)
 
   expect_equal(
     abs(colMeans(log_vol) - gold$log_volume) < matrixStats::colSds(log_vol),
