+++ R/generate.R
@@ -20,8 +20,7 @@
 #' `rlib_message_verbosity` is used to determine whether to show progress.
 #' @param allow_par `[logical(1)]`\cr `r lifecycle::badge("experimental")` If
 #' `TRUE`, run the nested sampling loop in parallel by breaking `nlive` into
-#' multiple samplers. Use [mirai::daemons()] to control how many samplers `x` is
-#' split into.
+#' multiple samplers.
 #'
 #' @returns `[ernest_run]`
 #'
