+++ R/ernest-package.R
@@ -12,3 +12,6 @@
 #' @useDynLib ernest, .registration = TRUE
 ## usethis namespace: end
 NULL
+
+# CRAN flags on parallelized variables
+utils::globalVariables(c("impl_", "lrps_"))
