+++ tests/testthat/_snaps/sampler-class.md
@@ -5,6 +5,6 @@
     Message
       Nested sampling run specification:
       * No. points: 500
-      * Sampling method: 25-step random walk sampling (acceptance target = 50.0%)
+      * Sampling method: 25-step random walk sampling (acceptance target = 50%)
       * Prior: uniform prior distribution with 2 dimensions (A and B)
 
