+++ NAMESPACE
@@ -106,8 +106,6 @@ importFrom(posterior,as_draws)
 importFrom(posterior,as_draws_matrix)
 importFrom(posterior,as_draws_rvars)
 importFrom(posterior,draws_of)
-importFrom(prettyunits,pretty_round)
-importFrom(prettyunits,pretty_signif)
 importFrom(tibble,tbl_sum)
 importFrom(tibble,tibble)
 importFrom(utils,head)
