+++ DESCRIPTION
@@ -20,11 +20,9 @@ Imports:
     lifecycle,
     matrixStats,
     posterior,
-    prettyunits,
     rlang (>= 1.1.0),
     tibble,
     uniformly,
-    utils,
     vctrs,
     withr
 Suggests: 
