### 2026-03-30 c8b8735
Minor optimization to picking copy-from point in LRPS.
