### 2026-03-31 e3ca3cc
New documentation for parallelization.
