+++ man/ernest-parallel.Rd
@@ -1,0 +1,106 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/parallel.R
+\name{ernest-parallel}
+\alias{ernest-parallel}
+\alias{parallelization}
+\title{Generate nested sampling runs in parallel.}
+\description{
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
+
+The cost of a nested sampling run depends on the distance between the prior
+and posterior distributions and on the number of live points. With
+\code{allow_par = TRUE}, \code{\link[=generate]{generate()}} uses \CRANpkg{mirai} to split the live set
+across workers, run sampling in parallel, then merge results before
+estimating model evidence.
+
+To run in parallel, daemons must be set with \code{\link[mirai:daemons]{mirai::daemons()}}. Otherwise,
+\code{\link[mirai:require_daemons]{mirai::require_daemons()}} throws an error when \code{allow_par = TRUE}.
+
+User-supplied \link{ernest_likelihood} and \link{ernest_prior} functions must be
+self-contained and must not depend on objects in the global environment.
+This keeps serialization predictable and avoids sending large, accidental
+dependencies to workers.
+}
+\section{Creating self-contained functions}{
+
+
+Self-contained likelihood and prior functions should follow these
+guidelines:
+\enumerate{
+\item Call package functions with explicit \code{::} namespaces, e.g.
+\code{extraDistr::qtnorm()}. Alternatively, call \code{library()} inside the function
+if you need to attach a package.
+\item Declare all data dependencies explicitly, similar to the pattern used in
+\link{create_likelihood}.
+\item Any helper functions (closures) called within a self-contained function
+must itself be self-contained.
+}
+}
+
+\section{Setting daemons}{
+
+
+How and where parallelisation occurs is determined by \code{\link[mirai:daemons]{mirai::daemons()}}.
+Daemons are persistent background processes that execute parallel
+computations locally or across a network.
+
+Daemons must be set before parallel execution. Otherwise, calling
+\code{generate(..., allow_par = TRUE)} throws an error.
+
+Usually, daemons are set once per session and can be left running while idle
+because they use minimal resources. The following sets up 6 local daemons:
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{mirai::daemons(6)
+}\if{html}{\out{</div>}}
+
+Function arguments:
+\itemize{
+\item \code{n}: the number of daemons to launch on your local machine, e.g.
+\code{mirai::daemons(6)}. As a rule of thumb, for maximum efficiency this should
+be (at most) one less than the number of cores on your machine, leaving one
+core for the main R process.
+\item \code{url} and \code{remote}: used to set up and launch daemons for distributed
+computing over the network. See \code{\link[mirai:daemons]{mirai::daemons()}} for more details.
+}
+
+Daemons persist for the duration of your session. To reset and shut them
+down:
+
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{mirai::daemons(0)
+}\if{html}{\out{</div>}}
+
+All daemons automatically terminate when your session ends. You do not need
+to explicitly terminate daemons in this instance, although it is still good
+practice to do so.
+}
+
+\examples{
+prior <- create_uniform_prior(lower = c(-1, -1), upper = 1)
+ll_fn <- function(x) -sum(x^2)
+sampler <- ernest_sampler(ll_fn, prior, nlive = 300)
+sampler
+
+# Initialise daemons
[…]
