+++ tests/testthat/test-algorithm.R
@@ -28,3 +28,27 @@ test_that("increasing min_logz reduces the iterations needed to converge", {
   )
   expect_gt(reference_run$niter, run_short$niter)
 })
+
+#' @srrstats {BS7.0, BS7.1} Parameter recovery for a normal distribution,
+#' without any additional information beyond the normal log-likelihood dist.
+test_that("Parameter recovery for a normal distribution", {
+  prior <- create_normal_prior(mean = c(0, 0))
+  log_l <- create_likelihood(
+    \(x) {
+      mvtnorm::dmvnorm(x, mean = c(0, 0), sigma = diag(2), log = TRUE)
+    }
+  )
+
+  sampler <- ernest_sampler(log_l, prior, nlive = 100, seed = 42)
+  run <- generate(sampler, max_iterations = 1000)
+  draws <- as_draws(run) |> posterior::resample_draws()
+  smry <- posterior::summarise_draws(
+    draws,
+    \(x) quantile(x, probs = c(0.05, 0.95))
+  )
+
+  expect_lte(-1.644854, smry[1, 2])
+  expect_gte(1.644854, smry[1, 3])
+  expect_lte(-1.644854, smry[2, 2])
+  expect_gte(1.644854, smry[2, 3])
+})
