### 2026-04-08 fce73a5
"ernest_rcrd" now is used directly for calculations without extracting fields.
