### 2026-04-08 91b224b
ernest_run now stores an ernest_rcrd rather than a 'weights' named list.
