+++ R/generate.R
@@ -41,20 +41,26 @@
 #'   transformed to the parameter space.
 #'   * `unit_cube`: `[matrix(double(), nsample, ndim)]` The samples in
 #'   the unit hypercube.
-#' * `weights`: `[list]` Additional information on the points generated by the
-#'   run:
-#'   * `id`: `[integer(nsample)]` The index of each point in the live set
-#'   for each sample.
-#'   * `evaluations`: `[integer(nsample)]` The number of likelihood
-#'   evaluations needed to replace each point.
 #'   * `log_weight`: `[double(nsample)]` The log-weight of each sample.
 #'   * `imp_weight`: `[double(nsample)]` The normalized importance weight
 #'   for each sample. This is given by normalizing `log_weight` with
 #'   `log_evidence`, then applying `exp` to the result so the weights sum to
 #'   one.
-#'   * `log_lik`: `[double(nsample)]` The log-likelihood for each sample.
-#'   * `birth_lik`: `[double(nsample)]` The log-likelihood of the point used
-#'   to create each sample.
+#' * `rcrd`: `[ernest_rcrd]` A subclass of [vctrs::new_rcrd] that stores
+#' sample-wise information on the run. This includes the following information
+#' which can be accessed with [[vctrs::field]]:
+#'   * `unit`: `[double(ndim)]` The location of the sample within the unit
+#'   hypercube.
+#'   * `id`: `[integer(nsample)]` The index of each sample within the live set.
+#'   * `nlive`: `[integer(nsample)]` The number of points within the live set
+#'   when this point was replaced.
+#'   * `evals`: `[integer(nsample)]` The number of likelihood evaluations needed
+#'   to replace this sample; equal to `0` if the sample is a member of the live
+#'   set at the end of the run.
+#'   * `log_lik`: `[double(nsample)]` The sample's log-likelihood.
+#'   * `birth_lik`: `[double(nsample)]` The log-likelihood of the point that was
+#'   used to create this sample; equal to `-Inf` if the sample was created when
+#'   the sampler was first compiled.
 #'
 #' @details
 #' At least one of `max_iterations`, `max_evaluations`, or `min_logz`

[…]
