### 2026-04-08 03c29f1
Dropping internal use of "x$samples" to store parameters. Also dropping "as_ernest_rcrd" helper in favour of directly accessing x$rcrd.
