+++ tests/testthat/test-rcrd-class.R
@@ -1,7 +1,7 @@
-test_that("as_ernest_rcrd works", {
+test_that("ernest_rcrd stores the expected information", {
   data(example_run)
-  x <- as_ernest_rcrd(example_run)
-  ref <- as_ernest_rcrd(example_run)
+  x <- example_run$rcrd
+  ref <- example_run$rcrd
   expect_equal(
     vctrs::fields(x),
     c("unit", "log_lik", "id", "nlive", "evals", "birth_lik")
@@ -10,8 +10,7 @@ test_that("as_ernest_rcrd works", {
   expect_equal(field(x, "id"), field(ref, "id"))
   expect_equal(field(x, "evals"), field(ref, "evals"))
   expect_equal(field(x, "birth_lik"), field(ref, "birth_lik"))
-  unit <- example_run$samples$unit_cube
-  dimnames(unit) <- NULL
+  unit <- field(example_run$rcrd, "unit")
   expect_equal(field(x, "unit"), unit)
   expect_identical(attr(x, "nvariables"), 3L)
   expect_snapshot(x)
@@ -19,27 +18,27 @@ test_that("as_ernest_rcrd works", {
 
 test_that("as.list ernest_rcrd works", {
   data(example_run)
-  x <- as_ernest_rcrd(example_run)
+  x <- example_run$rcrd
   list_x <- as.list(x)
   expect_named(
     list_x,
     c("unit", "log_lik", "id", "nlive", "evals", "birth_lik")
   )
-  expect_equal(list_x$log_lik, field(as_ernest_rcrd(example_run), "log_lik"))
+  expect_equal(list_x$log_lik, field(example_run$rcrd, "log_lik"))
 })
 
 test_that("ernest_rcrd orders draws", {
-  dead <- as_ernest_rcrd(example_run, keep_live = FALSE)
-  live <- tail(as_ernest_rcrd(example_run), 1000)
+  dead <- head(example_run$rcrd, -1000)
+  live <- tail(example_run$rcrd, 1000)
   scrambled_d <- sample(dead, size = length(dead))
   scrambled_l <- sample(live, size = length(live))
   all <- c(scrambled_d, scrambled_l)
   all <- sort(all)
   expect_equal(
     field(all, "log_lik"),
-    field(as_ernest_rcrd(example_run), "log_lik")
+    field(example_run$rcrd, "log_lik")
   )
-  expect_equal(field(all, "unit"), field(as_ernest_rcrd(example_run), "unit"))
+  expect_equal(field(all, "unit"), field(example_run$rcrd, "unit"))
   expect_identical(attr(all, "nvariables"), 3L)
 })
 
@@ -61,7 +60,7 @@ test_that("ernest_rcrd reports dimensional mismatch", {
     max_iterations = 1000
   )
   expect_error(
-    c(as_ernest_rcrd(run2D), as_ernest_rcrd(run3D)),
+    c(run2D$rcrd, run3D$rcrd),
     "`variables` attribute must match."
   )
 })
