+++ tests/testthat/test-edgecases.R
@@ -107,7 +107,7 @@ test_that("Missing values in the log-likelihood", {
   )
 
   run <- generate(quiet_na_sampler)
-  expect_false(anyNA(vctrs::field(as_ernest_rcrd(run), "log_lik")))
+  expect_false(anyNA(vctrs::field(run$rcrd, "log_lik")))
 })
 
 #' Special tests: Perfectly flat and nearly flat likelihoods
@@ -143,5 +143,5 @@ test_that("Ernest halts and warns when ll becomes flat during a run", {
     ),
     "Stopping run due to a likelihood plateau at 0"
   )
-  expect_all_equal(tail(vctrs::field(as_ernest_rcrd(run), "log_lik"), 100), 0)
+  expect_all_equal(tail(vctrs::field(run$rcrd, "log_lik"), 100), 0)
 })

[…]
