+++ NAMESPACE
@@ -57,6 +57,7 @@ S3method(vec_ptype2,list.ernest_rcrd)
 S3method(vec_ptype_abbr,ernest_rcrd)
 S3method(vec_ptype_full,ernest_rcrd)
 S3method(visualize,ernest_run)
+S3method(weights,ernest_run)
 export(as_draws)
 export(as_draws_matrix)
 export(as_draws_rvars)

[…]
