### 2026-04-08 960b039
Moved tests so `use_test/use_r` conventions are followed by the whole package.
