### 2026-04-09 e966630
New calculate method for ndraws = 0.
