+++ R/calculate.R
@@ -6,9 +6,9 @@
 #'
 #' @param x [[ernest_run]]\cr Results from a nested sampling run.
 #' @param ndraws `[integer(1)]`\cr The number of log-volume sequences to
-#' simulate.
-#' If equal to zero, no simulations will be made, and a one draw vector of
-#' log-volumes are produced from the estimates contained in `x`.
+#' simulate. If equal to zero, log-volume simulation is skipped and error in
+#' the log-evidence estimates is approximated with analytical error
+#' estimates.
 #' @inheritParams rlang::args_dots_empty
 #'
 #' @returns [[tibble::tibble()]] with class `ernest_estimate`.
@@ -16,17 +16,19 @@
 #' The iterative estimates from the nested sampling run. Contains the following
 #' columns:
 #' * `log_lik`: `[[double()]]` The log-likelihood of the model.
-#' * `log_volume`: `[[double() or rvar]]` The log-volume of the prior space.
-#' * `log_weight`: `[[double() or rvar]]` The log weights of the points in the
+#' * `log_volume`: `[[posterior::rvar()]]` The log-volume of the prior space.
+#' * `log_weight`: `[[posterior::rvar()]]` The log weights of the points in the
 #' live set.
-#' * `log_evidence`: `[[double() or rvar]]` The log-evidence of the model.
+#' * `log_evidence`: `[[posterior::rvar()]]` The log-evidence of the model.
 #'
-#' If `ndraws = 0`, columns are retunred as doubles. Else, they are returned
-#' as [[posterior::rvar()]] objects, with `ndraws` rows and the same number of
-#' columns as the original `log_volume` and `log_weight` matrices.
+#' If `ndraws > 0`, `log_volume`, `log_weight`, and `log_evidence` each contain
+#' `ndraws` simulated draws per iteration.
 #'
-#' If `ndraws = 0`,  an additional column is included:
-#' * `log_evidence_err`: `[[double()]]` The standard error of the log-evidence.
+#' If `ndraws = 0`, `log_volume` and `log_weight` contain a single
+#' deterministic draw per iteration, and `log_evidence` contains draws
+#' from a normal approximation based on analytical variance estimates (see
+#' the package vignetttes for more information). The  number of draws is
+#' controlled with getOption("posterior.rvar_ndraws"), with a default of 1000.
 #'
 #' @references Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
 #' Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
@@ -57,33 +59,29 @@ calculate.ernest_run <- function(x, ndraws = 1000L, ...) {
   log_volume <- get_log_vol(x_rcrd, ndraws = ndraws)
   log_weight <- get_log_w(log_lik, log_volume)
 
-  result <- if (ndraws == 0) {
-    list(
-      "log_lik" = log_lik,
-      "log_volume" = drop(log_volume),
-      "log_weight" = drop(log_weight$log_weight),
-      "log_evidence" = drop(log_weight$log_evidence),
-      "log_evidence_err" = {
-        information <- get_information(
-          log_lik,
-          log_volume,
-          drop(log_weight$log_evidence)
-        )
-        sqrt(get_log_zvar(information, log_volume))
-      }
+  log_z <- if (ndraws == 0) {
+    information <- get_information(log_lik, log_volume, log_weight$log_evidence)
+    log_evidence_var <- get_log_zvar(information, log_volume)
[…]
