### 2026-04-10 4cbeee2
Fixes to dimensionality errors on calculate.
