+++ tests/testthat/_snaps/calculate.md
@@ -5,18 +5,39 @@
     Output
       # <ernest_estimate>:  10359 niter.
       # Uncertainty source: Normally-Approximated Analytical Estimates (1000 draws)
-         log_lik  log_volume log_weight log_evidence
-           <dbl>  <rvar[1d]> <rvar[1d]>   <rvar[1d]>
-       1   -137.  -5.2 ± 3.1  -143 ± NA     -143 ± 0
-       2   -132.  -5.2 ± 3.1  -139 ± NA     -139 ± 0
-       3   -130.  -5.2 ± 3.1  -137 ± NA     -137 ± 0
-       4   -130.  -5.2 ± 3.1  -137 ± NA     -136 ± 0
-       5   -129.  -5.2 ± 3.1  -136 ± NA     -136 ± 0
-       6   -127.  -5.2 ± 3.1  -134 ± NA     -134 ± 0
-       7   -124.  -5.2 ± 3.1  -131 ± NA     -131 ± 0
-       8   -123.  -5.2 ± 3.1  -130 ± NA     -130 ± 0
-       9   -123.  -5.2 ± 3.1  -130 ± NA     -129 ± 0
-      10   -122.  -5.2 ± 3.1  -129 ± NA     -129 ± 0
+         log_lik   log_volume log_weight log_evidence
+           <dbl>   <rvar[1d]> <rvar[1d]>   <rvar[1d]>
+       1   -137.  -0.001 ± NA  -143 ± NA     -143 ± 0
+       2   -132.  -0.002 ± NA  -139 ± NA     -139 ± 0
+       3   -130.  -0.003 ± NA  -137 ± NA     -137 ± 0
+       4   -130.  -0.004 ± NA  -137 ± NA     -136 ± 0
+       5   -129.  -0.005 ± NA  -136 ± NA     -136 ± 0
+       6   -127.  -0.006 ± NA  -134 ± NA     -134 ± 0
+       7   -124.  -0.007 ± NA  -131 ± NA     -131 ± 0
+       8   -123.  -0.008 ± NA  -130 ± NA     -130 ± 0
+       9   -123.  -0.009 ± NA  -130 ± NA     -129 ± 0
+      10   -122.  -0.010 ± NA  -129 ± NA     -129 ± 0
+      # i 10,349 more rows
+
+# calculate / works when ndraws = 1 (default)
+
+    Code
+      calc
+    Output
+      # <ernest_estimate>:  10359 niter.
+      # Uncertainty source: Simulated Log-Volumes (1 draws)
+         log_lik    log_volume log_weight log_evidence
+           <dbl>    <rvar[1d]> <rvar[1d]>   <rvar[1d]>
+       1   -137.  -0.0061 ± NA  -142 ± NA    -142 ± NA
+       2   -132.  -0.0088 ± NA  -138 ± NA    -138 ± NA
+       3   -130.  -0.0116 ± NA  -137 ± NA    -136 ± NA
+       4   -130.  -0.0121 ± NA  -137 ± NA    -136 ± NA
+       5   -129.  -0.0134 ± NA  -137 ± NA    -136 ± NA
+       6   -127.  -0.0134 ± NA  -134 ± NA    -134 ± NA
+       7   -124.  -0.0148 ± NA  -132 ± NA    -132 ± NA
+       8   -123.  -0.0149 ± NA  -131 ± NA    -131 ± NA
+       9   -123.  -0.0156 ± NA  -131 ± NA    -130 ± NA
+      10   -122.  -0.0158 ± NA  -131 ± NA    -130 ± NA
       # i 10,349 more rows
 
[…]
