### 2026-04-10 c9ca90c
New plot methods based on calculate.
