+++ R/plot.R
@@ -78,23 +78,120 @@ plot.ernest_estimate <- function(
   print(autoplot(x, which, ...)) # nocov
 }
 
-#' @inheritParams calculate.ernest_run
-#'
-#' @rdname plot.ernest_run
+#' @rdname plot.ernest_estimate
 #' @export
 plot.ernest_run <- function(
   x,
   which = c("evidence", "weight", "likelihood"),
-  ndraws = NULL,
   ...
 ) {
   which <- check_plot_which(which)
-  if (!is.null(ndraws)) {
-    x <- calculate(x, ndraws = ndraws)
-  }
-  print(autoplot(x, which, ...)) # nocov
+  withr::with_preserve_seed(
+    obj <- calculate(x, ndraws = 0)
+  )
+  print(autoplot(obj, which, ...))
 }
 
+#' @rdname plot.ernest
+#' @export
+summary.ernest_estimate <- function(
+  object,
+  which = c("evidence", "weight", "likelihood"),
+  n = 512,
+  ...
+) {
+  check_dots_empty()
+  which <- arg_match(which, multiple = TRUE)
+  check_number_whole(n, min = 2)
+
+  if (any(c("evidence", "weight") %in% which)) {
+    check_installed("ggdist", "to summarize evidence and weight diagnostics")
+  }
+
+  log_volume <- posterior::draws_of(object$log_volume)
+  log_evidence <- posterior::draws_of(object$log_evidence)
+  log_weight <- posterior::draws_of(object$log_weight)
+  weight <- if (nrow(log_weight) == 1) {
+    log_z <- matrixStats::logSumExp(drop(log_weight))
+    exp(log_weight - log_z)
+  } else {
+    exp(sweep(
+      log_weight,
+      1,
+      log_evidence[, ncol(log_evidence)],
+      FUN = "-"
+    ))
+  }
+
+  lik <- exp(object$log_lik - max(object$log_lik))
+  vol_rng <- ggdist::density_bounded(
+    log_volume,
+    bounds = c(NA, 0),
+    range_only = TRUE
+  )$x
+  knots <- seq(vol_rng[1], vol_rng[2], length.out = n)
+  print(length(knots))
+
+  safe_curve <- \(x, y) {
+    if (!is.matrix(drop(y))) {
+      vctrs::data_frame(
+        "x" = x,
+        ".value" = y
+      )
+    } else {
+      ggdist::curve_interval(
+        vctrs::data_frame(
+          "x" = x,
+          ".value" = posterior::rvar(t(y))
+        ),
+        .along = x,
+        .width = ggdist::interval_widths(3)
+      )
+    }
+  }
+
+  list(
+    "evidence" = if ("evidence" %in% which) {
+      yout <- vapply(
+        seq_len(nrow(log_evidence)),
+        \(i) {
+          stats::approx(
+            log_volume[min(i, nrow(log_volume)), ],
+            exp(log_evidence[i, ]),
+            xout = knots,
+            rule = 2
+          )$y
+        },
+        double(n)
+      )
+      safe_curve(knots, yout)
+    },
+    "weight" = if ("weight" %in% which) {
+      yout <- vapply(
+        seq_len(nrow(weight)),
+        \(i) {
+          dens <- ggdist::density_bounded(
+            x = log_volume[i, ],
+            weights = weight[i, ],
+            bounds = c(NA, 0)
+          )
+          stats::approx(dens$x, dens$y, xout = knots, rule = 2)$y
+        },
+        double(n)
+      )
+      safe_curve(knots, yout)
+    },
+    "likelihood" = if ("likelihood" %in% which) {
+      lik_app <- approx(
+        colMeans(log_volume),
+        lik,
+        xout = knots,
+        rule = 2
+      )
[…]
