+++ NAMESPACE
@@ -6,7 +6,6 @@ S3method(as_draws,ernest_run)
 S3method(as_draws_matrix,ernest_run)
 S3method(as_draws_rvars,ernest_run)
 S3method(autoplot,ernest_estimate)
-S3method(autoplot,ernest_run)
 S3method(calculate,ernest_run)
 S3method(compile,ernest_run)
 S3method(compile,ernest_sampler)
@@ -39,6 +38,7 @@ S3method(propose,slice_rectangle)
 S3method(propose,unif_cube)
 S3method(propose,unif_ellipsoid)
 S3method(stats::weights,ernest_run)
+S3method(summary,ernest_estimate)
 S3method(summary,ernest_run)
 S3method(tbl_sum,ernest_estimate)
 S3method(tbl_sum,ernest_resample)
@@ -95,7 +95,6 @@ importFrom(ggplot2,autoplot)
 importFrom(ggplot2,facet_grid)
 importFrom(ggplot2,geom_line)
 importFrom(ggplot2,geom_point)
-importFrom(ggplot2,geom_ribbon)
 importFrom(ggplot2,geom_vline)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,scale_colour_distiller)
