### 2026-04-11 9f39f5a
Plots for ndraws = 0 are now reliable.
