### 2026-04-13 8c80e2b
Preserve seed + distributional for plotting ndraws = 0.
