### 2026-04-13 1272057
Reorganized plot method.
