+++ R/plot.R
@@ -73,7 +73,6 @@ plot.ernest_estimate <- function(
   which = c("evidence", "weight", "likelihood"),
   ...
 ) {
-  check_dots_empty()
   which <- check_plot_which(which)
   print(autoplot(x, which, ...)) # nocov
 }
@@ -98,12 +97,14 @@ summary.ernest_estimate <- function(
   object,
   which = c("evidence", "weight", "likelihood"),
   n = 512,
+  width = 0.5,
   ...
 ) {
   check_dots_empty()
   which <- arg_match(which, multiple = TRUE)
+  width <- vctrs::vec_cast(width, to = double())
+  nwidth <- length(width)
   check_number_whole(n, min = 2)
-
   if (any(c("evidence", "weight") %in% which)) {
     check_installed("ggdist", "to summarize evidence and weight diagnostics")
   }
@@ -124,14 +125,9 @@ summary.ernest_estimate <- function(
   }
 
   lik <- exp(object$log_lik - max(object$log_lik))
-  vol_rng <- ggdist::density_bounded(
-    log_volume,
-    bounds = c(NA, 0),
-    range_only = TRUE
-  )$x
-  knots <- seq(vol_rng[1], vol_rng[2], length.out = n)
-  print(length(knots))
+  knots <- seq(mean(log_volume[, ncol(log_volume)]), 0, length.out = n)
 
+  # Safely summarize curves with or without uncertainty intervals.
   safe_curve <- \(x, y) {
     if (!is.matrix(drop(y))) {
       vctrs::data_frame(
@@ -145,81 +141,85 @@ summary.ernest_estimate <- function(
           ".value" = posterior::rvar(t(y))
         ),
         .along = x,
-        .width = ggdist::interval_widths(3)
+        .width = width
       )
     }
   }
 
-  list(
-    "evidence" = if ("evidence" %in% which) {
-      if (nrow(log_volume) == 1) {
-        z_app <- approx(
-          log_volume[1, ],
-          colMeans(log_evidence),
-          xout = knots,
-          rule = 2
-        )
-        zsd_app <- approx(
-          log_volume[1, ],
-          sqrt(attr(object, "log_z_var")),
-          xout = knots,
-          rule = 2
-        )
-        tmp <- vctrs::vec_expand_grid(
-          "log_z" = vctrs::data_frame(
-            "x" = z_app$x,
-            "m" = z_app$y,
-            "sd" = zsd_app$y
-          ),
-          ".width" = ggdist::interval_widths(3)
-        )
-        vctrs::data_frame(
-          "x" = tmp$log_z$x,
-          ".value" = exp(tmp$log_z$m),
-          ".width" = tmp$.width,
-          ".lower" = exp(tmp$log_z$m - tmp$log_z$sd * tmp$.width),
-          ".upper" = exp(tmp$log_z$m + tmp$log_z$sd * tmp$.width)
-        )
-      } else {
-        yout <- vapply(
-          seq_len(nrow(log_evidence)),
-          \(i) {
-            stats::approx(
-              log_volume[i, ],
-              exp(log_evidence[i, ]),
-              xout = knots,
-              rule = 2
-            )$y
-          },
-          double(n)
-        )
-        safe_curve(knots, yout)
-      }
-    },
-    "weight" = if ("weight" %in% which) {
-      yout <- vapply(
-        seq_len(nrow(weight)),
-        \(i) {
-          dens <- ggdist::density_bounded(
-            x = log_volume[i, ],
-            weights = weight[i, ],
-            bounds = c(NA, 0)
-          )
-          stats::approx(dens$x, dens$y, xout = knots, rule = 2)$y
-        },
[…]
