+++ tests/testthat/test-plot.R
@@ -1,112 +1,70 @@
-test_that("plot.ernest_estimate validates input arguments", {
+test_that("summary.ernest_estimate works when ndraws = 0", {
+  data(example_run)
+  obj <- calculate(example_run, ndraws = 0)
+  base <- c("x", ".value")
+  out <- summary(
+    obj,
+    which = c("evidence", "weight", "likelihood"),
+    n = 512,
+    width = c(0.5, 0.8, 0.95)
+  )
+
+  expect_snapshot_value(out$evidence, tolerance = 1e-3, style = "json2")
+  expect_snapshot_value(out$weight, tolerance = 1e-3, style = "json2")
+  expect_snapshot_value(out$likelihood, tolerance = 1e-3, style = "json2")
+})
+
+test_that("summary.ernest_estimate works when ndraws = 100", {
+  data(example_run)
+  withr::local_seed(42)
+  obj <- calculate(example_run, ndraws = 100)
+  base <- c("x", ".value")
+  curve <- c(
+    c(".lower", ".upper", ".actual_width"),
+    c(".width", ".point", ".interval")
+  )
+  out <- summary(
+    obj,
+    which = c("evidence", "weight", "likelihood"),
+    n = 512,
+    width = c(0.5, 0.8, 0.95)
+  )
+
+  expect_named(out, c("evidence", "weight", "likelihood"))
+  expect_named(out$evidence, c(base, curve))
+  expect_named(out$weight, c(base, curve))
+  expect_named(out$likelihood, base)
+})
+
+test_that("plot an ernest_estimate object", {
   set.seed(42)
-  data(example_run)
-  est <- calculate(example_run, ndraws = 10)
-
-  expect_error(
-    plot(est, which = "not_a_plot"),
-    "must be one of"
-  )
-  expect_error(
-    plot(est, which = c("evidence", "not_a_plot")),
-    "must be one of"
-  )
-  expect_error(
-    plot(est, which = character()),
-    "At least one plot type must be specified in `which`."
-  )
-})
-
-test_that("plot.ernest_run validates input arguments", {
-  data(example_run)
-  expect_error(
-    plot(example_run, which = "not_a_plot"),
-    "must be one of"
-  )
-})
-
-skip_plot_snapshot()
-describe("plotting an ernest_estimate object", {
-  skip()
-  set.seed(42)
-  calc_100 <- calculate(example_run, ndraws = 100)
-
-  it("plots evidence", {
-    vdiffr::expect_doppelganger(
-      "evidence ndraws = 100",
-      plot(calc_100, which = "evidence")
-    )
-  })
-
-  it("plots weight", {
-    vdiffr::expect_doppelganger(
-      "weight ndraws = 100",
-      plot(calc_100, which = "weight")
-    )
-  })
-
-  it("plots likelihood", {
-    vdiffr::expect_doppelganger(
-      "likelihood ndraws = 100",
-      plot(calc_100, which = "likelihood")
-    )
-  })
-
-  it("plots everything", {
-    vdiffr::expect_doppelganger(
-      "all ndraws = 1",
-      plot(calculate(example_run, ndraws = 1))
-    )
-
-    vdiffr::expect_doppelganger(
-      "all ndraws = 100",
-      plot(calc_100)
-    )
-  })
-})
-
-describe("plotting an ernest_run object", {
-  data(example_run)
-  it("plots evidence", {
-    vdiffr::expect_doppelganger(
-      "evidence ernest_run",
-      plot(example_run, which = "evidence")
-    )
-  })
-
-  it("plots weights", {
-    vdiffr::expect_doppelganger(
-      "weight ernest_run",
-      plot(example_run, which = "weight")
[…]
