### 2026-04-13 c002220
New documentation for plot.
