+++ R/plot.R
@@ -1,47 +1,44 @@
 #' Plot diagnostics from nested sampling results
 #'
-#' Visualizes key diagnostics from nested sampling outputs, including the
-#' normalized likelihood, posterior weights, and evidence as functions of
+#' Visualizes key diagnostics from nested sampling outputs as functions of
 #' log-prior volume.
 #'
 #' @param x [[ernest_run]] or [[ernest_estimate]]\cr An object containing
 #' results from nested sampling.
-#' @param which `[character()]`\cr Choose which plots to display. Must be one
-#' or more of `"evidence"`, `"weight"`, and `"likelihood"`.
+#' @param which `[character()]`\cr One or more diagnostics to display. Must be
+#' any of `"evidence"`, `"weight"`, and `"likelihood"`.
+#' @param n `[integer(1)]`\cr Number of evaluation points along the
+#' log-volume axis used to summarize each curve.
 #' @inheritParams rlang::args_dots_empty
 #'
-#' @returns `[invisible(x)]`. A `ggplot2::ggplot()` object is printed as a
-#' side effect.
+#' @returns
+#' For `plot`, the `ggplot` object, invisibly. It is also printed as a side
+#' effect.
+#' For `summary`, a `list` with possible elements `evidence`, `weight`,
+#' and `likelihood`. Each element is a data frame.
 #'
 #' @details
-#' The plotting methods are structured around internal `autoplot_*` helpers
-#' (`autoplot_evidence()`, `autoplot_weight()`, and `autoplot_likelihood()`).
-#' Use `which` to select one or more diagnostics:
+#' `plot()` is a visualization wrapper around [summary.ernest_estimate()],
+#' followed by `autoplot()`. Use `which` to select diagnostics:
 #'
-#' * `which = "evidence"`: Plots the estimated marginal
-#' likelihood (evidence) as a function of log-prior volume, with uncertainty
-#' intervals. Peaks in this plot indicate regions of prior volume that
-#' contribute most to the evidence estimate.
-#' * `which = "weight"`: Shows the distribution of
-#' posterior mass across log-prior volume. This plot highlights which regions
-#' of the prior volume contain the most posterior probability, helping to
-#' identify where the sampler concentrated its effort.
-#' * `which = "likelihood"`: Displays the normalized
-#' likelihood as a function of log-prior volume. Smoothness in this plot
-#' indicates effective likelihood-restricted prior sampling, while
-#' irregularities may suggest sampling difficulties or, in some cases,
-#' misspecified likelihood functions.
+#' * `which = "evidence"`: Estimated marginal likelihood as a function of
+#' log-prior volume.
+#' * `which = "weight"`: Posterior mass concentration across log-prior volume.
+#' * `which = "likelihood"`: Normalized likelihood across log-prior volume.
 #'
-#' If `x` is an `ernest_run`, plotting uses deterministic integral estimates
-#' (`calculate(x, ndraws = 0)`). For `which = "evidence"`, uncertainty bands
-#' are derived from the analytic standard error stored in
-#' `x$log_evidence_err`.
+#' If `x` is an `ernest_run`, plotting first computes
+#' `calculate(x, ndraws = 0)`. In this mode, `log_volume` and `log_weight` are
+#' deterministic, and evidence uncertainty comes from the analytical normal
+#' approximation generated from the original sampling run.
 #'
-#' If `x` is an `ernest_estimate` (or if `ndraws` is specified), the plots
-#' are based on simulated log-volume trajectories. In this case,
[…]
