+++ man/plot.ernest.Rd
@@ -1,0 +1,92 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plot.R
+\name{plot.ernest_estimate}
+\alias{plot.ernest_estimate}
+\alias{plot.ernest_run}
+\alias{summary.ernest_estimate}
+\title{Plot diagnostics from nested sampling results}
+\usage{
+\method{plot}{ernest_estimate}(x, which = c("evidence", "weight", "likelihood"), n = 512, ...)
+
+\method{plot}{ernest_run}(x, which = c("evidence", "weight", "likelihood"), n = 512, ...)
+
+\method{summary}{ernest_estimate}(
+  object,
+  which = c("evidence", "weight", "likelihood"),
+  n = 512,
+  width = NULL,
+  ...
+)
+}
+\arguments{
+\item{x}{[\link{ernest_run}] or [\link{ernest_estimate}]\cr An object containing
+results from nested sampling.}
+
+\item{which}{\verb{[character()]}\cr One or more diagnostics to summarize. Must be
+any of \code{"evidence"}, \code{"weight"}, and \code{"likelihood"}.}
+
+\item{n}{\verb{[integer(1)]}\cr Number of evaluation points along the
+log-volume axis used to summarize each curve.}
+
+\item{...}{These dots are for future extensions and must be empty.}
+
+\item{object}{[\link{ernest_estimate}]\cr Output from \code{\link[=calculate]{calculate()}}.}
+
+\item{width}{\verb{[double()]}\cr Vector of probabilities to use that determine
+the widths of the resulting intervals, as in [\link[ggdist:curve_interval]{ggdist::curve_interval}].
+Defaults to three widths roughly correspoding to +/- 1, 2, and 3 SD.}
+}
+\value{
+For \code{plot}, the \code{ggplot} object, invisibly. It is also printed as a side
+effect.
+For \code{summary}, a \code{list} with possible elements \code{evidence}, \code{weight},
+and \code{likelihood}. Each element is a data frame.
+}
+\description{
+Visualizes key diagnostics from nested sampling outputs as functions of
+log-prior volume.
+}
+\details{
+\code{plot()} is a visualization wrapper around \code{\link[=summary.ernest_estimate]{summary.ernest_estimate()}},
+followed by \code{autoplot()}. Use \code{which} to select diagnostics:
+\itemize{
+\item \code{which = "evidence"}: Estimated marginal likelihood as a function of
+log-prior volume.
+\item \code{which = "weight"}: Posterior mass concentration across log-prior volume.
+\item \code{which = "likelihood"}: Normalized likelihood across log-prior volume.
+}
+
+If \code{x} is an \code{ernest_run}, plotting first computes
+\code{calculate(x, ndraws = 0)}. In this mode, \code{log_volume} and \code{log_weight} are
+deterministic, and evidence uncertainty comes from the analytical normal
+approximation generated from the original sampling run.
+
+If \code{x} is an \code{ernest_estimate} generated with \code{ndraws > 0}, diagnostics are
+summarized over simulated log-volume trajectories. For these simulated
+estimates, uncertainty bands for \code{evidence} and \code{weight} are computed as
+interval summaries on interpolated curves.
+
+To get the underlying data frames used for plotting, use
+\code{summary.ernest_estimate()}. This is useful when you want full control over
+plotting.
+}
+\note{
+Plotting multiple diagnostics with \code{which} requires \CRANpkg{patchwork}.
+Plotting \code{evidence} or \code{weight} diagnostics requires \CRANpkg{ggdist}.
+}
+\examples{
+# Plot diagnostics from a run (analytical uncertainty for evidence).
+data(example_run)
[…]

[…]
