### 2026-04-13 08db0eb
Fixed a bug with birth_lik writing during a run.
