### 2026-04-14 2139c7b
Demoted several packages to "Suggests". Cleaned up vctrs imports.
