+++ R/crosscheck.R
@@ -1,0 +1,243 @@
+#' Visually examine the sampling behaviour of a nested sampling run
+#'
+#' @param x [[ernest_run]]\cr An 'ernest_run' object.
+#' @param size `[[integer(1)]]`\cr The number of points to include in each
+#' split. If `NULL`, this defaults to `x$nlive`. If larger than `x$niter`, the
+#' run is not split.
+#' @param conf.level `[[double(1)]]`\cr Confidence level for the plotted
+#' confidence intervals.
+#' @param data_only `[[logical(1)]]`\cr Whether to automatically plot the
+#' results. If `FALSE`, the plotted data is instead returned to the user.
+#'
+#' @rdname crosscheck_ernest
+#' @importFrom ggplot2 geom_step
+#' @export
+crosscheck_plot <- function(
+  x,
+  size = NULL,
+  conf.level = 0.95, # nolint
+  data_only = FALSE,
+  ...
+) {
+  check_class(x, "ernest_run")
+  check_number_decimal(conf.level)
+  if (conf.level <= 0.5 || conf.level >= 1) {
+    stop_input_type(conf.level, "a single number between 0.5 and 1")
+  }
+  check_bool(data_only)
+
+  splits <- crosscheck_prep(x, size)
+  null <- calculate_ecdf_ci(length(splits[[1]]), x$nlive, 1 - conf.level)
+  cumsums <- lapply(
+    splits,
+    \(x, nlive) {
+      data_frame0(
+        !!!null,
+        ".y" = c(0, cumsum(tabulate(x, nbins = nlive)))
+      )
+    },
+    nlive = x$nlive
+  )
+  names(cumsums) <- names(splits)
+  cdf <- vctrs::vec_rbind(!!!cumsums, .names_to = "idx")
+  if (data_only) {
+    return(cdf)
+  }
+  ggplot(cdf, aes(x = .data$.x)) +
+    geom_step(
+      aes(
+        y = .data$.y / .data$.n - .data$.p,
+        group = factor(.data$idx),
+        colour = factor(.data$idx)
+      )
+    ) +
+    geom_step(
+      aes(y = .data$.lower / .data$.n - .data$.p),
+      colour = "gray60"
+    ) +
+    geom_step(
+      aes(y = .data$.upper / .data$.n - .data$.p),
+      colour = "gray60"
+    ) +
+    ggplot2::annotate("segment", x = 0, y = 0, xend = x$nlive) +
+    ggplot2::labs(
+      x = "Insertion Index",
+      y = bquote(hat(F[n](k)) - F[n](k)),
+      colour = "Iterations"
+    )
+}
+
+#' @param type `[[character(1)]]`\cr The type of test statistic to use for the
+#' uniformity test. One of "W2" (Cramer-von Mises), "A2" (Anderson-Darling),
+#' "ks" (Kolmogorov-Smirnov). Tests are not performed if `type` is `NA`.
+#' @importFrom stats p.adjust.methods
+#'
+#' @rdname crosscheck_ernest
+#' @export
+crosscheck_tests <- function(
+  x,
+  size = NULL,
+  type = c("W2", "A2", "ks"),
+  p.adjust = p.adjust.methods,
+  ...
+) {
+  check_class(x, "ernest_run")
+  check_installed("dgof", "to perform uniformity tests")
+  y_ecdf <- stats::ecdf(1:x$nlive)
+  test_fn <- switch(
+    arg_match(type),
+    "W2" = \(x) dgof::cvm.test(x = x, y = y_ecdf, type = "W2", ...),
+    "A2" = \(x) dgof::cvm.test(x = x, y = y_ecdf, type = "A2", ...),
[…]
