### 2026-04-24 772998f
CUSUM plots for diagnostics!
