### 2026-04-29 79668bb
Revert CUSUM plot functionality.
This reverts commit 772998f07203615836b0e8faf3193c0bee5a906f.
