+++ R/crosscheck.R
@@ -1,202 +1,247 @@
-#' Visually diagnose sampling issues in an `ernest_run`
+#' Visually examine the sampling behaviour of a nested sampling run
 #'
-#' Issues in sampling caused by poor LRPS behaviour or challenging features
-#' in the model's likelihood may confound nested sampling results. To identify
-#' these issues, this plot tracks the insertion indicies of new points as they
-#' are inserted into the live set, calculates a rolling rank sum statistic,
-#' then creates a CUSUM plot highlighting potentially problematic sampling
-#' iterations.
+#' @param x [[ernest_run]]\cr An 'ernest_run' object.
+#' @param size `[[integer(1)]]`\cr The number of points to include in each
+#' split. If `NULL`, this defaults to `x$nlive`. If larger than `x$niter`, the
+#' run is not split.
+#' @param conf.level `[[double(1)]]`\cr Confidence level for the plotted
+#' confidence intervals.
+#' @param data_only `[[logical(1)]]`\cr Whether to automatically plot the
+#' results. If `FALSE`, the plotted data is instead returned to the user.
 #'
-#' @param run [[ernest_run]]\cr Results from a nested sampling run.
-#' @param width `[integer(1)]`\cr Width of the rolling window used to compute
-#' the rank-sum process. Defaults to `run$nlive` when left `NULL`.
-#' @param omega `[double(1)]`\cr Reference value used to create the cusum chart.
-#' Larger values make the chart less sensitive to small shifts in the rank sum
-#' of the insertions.
-#' @param h `[double(1)]`\cr Decision interval for out-of-bounds detection.
-#' Iterations are flagged when the upper CUSUM exceeds `h` or the lower CUSUM
-#' drops below `-h`.
-#' @param plot `[logical(1)]`\cr If `TRUE`, returns a CUSUM `ggplot` object. If
-#' `FALSE`, returns the tabular CUSUM values used to make the plot.
-#'
-#' @returns
-#' If `plot = TRUE`, a `ggplot2::ggplot()` object.
-#'
-#' If `plot = FALSE`, a tibble with columns:
-#' * `.iter`: `[integer()]` CUSUM iteration index.
-#' * `x`: `[double()]` Rolling rank-sum statistic.
-#' * `z`: `[double()]` Standardized rank-sum statistic.
-#' * `.lower`: `[double()]` Lower tabular CUSUM sequence.
-#' * `.upper`: `[double()]` Upper tabular CUSUM sequence.
-#' * `OOB`: `[logical()]` Indicator for whether either CUSUM is outside the
-#' decision interval [-h, h].
-#'
-#' @details
-#' The diagnostic follows these steps:
-#' * Compute insertion indices from the run's record of live-point
-#' replacements.
-#' * Build a rolling rank sum of insertion ranks using `width`.
-#' * Standardize this process using the expected mean and standard deviation
-#' under ideal insertion behavior.
-#' * Apply a one-sided tabular CUSUM to detect sustained positive or negative
-#' departures from expectation.
-#'
-#' Out-of-bounds (`OOB`) points indicate iterations where insertion behavior may
-#' be inconsistent with well-mixed sampling in the likelihood-restricted prior.
-#'
-#' @seealso
[…]
