+++ NAMESPACE
@@ -60,7 +60,6 @@ export(as_draws)
 export(as_draws_matrix)
 export(as_draws_rvars)
 export(calculate)
-export(calculate_cusum)
 export(compile)
 export(create_likelihood)
 export(create_normal_prior)

[…]
