### 2026-05-03 #730 @maurolepore
Apologies for the delay since your last check-in.

I see both reviews are in (I formally recorded that immediately above) and it seems that you've addressed the feedback in your repo.

The next step is for you to please post your response on this thread addressing each reviewer's points ([example](https://github.com/ropensci/software-review/issues/160#issuecomment-355043656), [guidelines](https://devguide.ropensci.org/softwarereview_author.html)).

Thanks for your patience and for sharing your work with the rOpenSci community.
