### 2026-05-05 8233c18
New bootstrap method, helpers for accessing common features from ernest_rcrd and ernest_run.
