+++ tests/testthat/test-boot.R
@@ -1,46 +1,39 @@
 withr::local_seed(42)
 data("example_run")
 
-test_that("learn rejects bad parameters", {
+test_that("generate_bootstraps input validation", {
+  expect_error(generate_bootstraps(example_run, times = -1))
+  expect_error(generate_bootstraps(example_run, times = 1.5))
   expect_error(
-    learn(example_run, times = 0),
-    "must be a whole number larger than or equal to 1"
-  )
-  expect_error(
-    learn(example_run, include_weights = "yes"),
-    "must be `TRUE` or `FALSE`"
-  )
-  expect_error(
-    learn(example_run, units = "bad"),
+    generate_bootstraps(example_run, units = "bad"),
     'must be one of "original" or "unit_cube"'
   )
 })
 
-test_that("learn returns expected structure", {
-  res <- learn(example_run)
+test_that("generate_bootstraps returns expected structure", {
+  res <- generate_bootstraps(example_run, times = 100)
 
-  expect_s3_class(res, c("ernest_resample", "tbl_df"))
+  expect_s3_class(res, "tbl_df")
   expect_equal(nrow(res), 100)
-  expect_type(res$log_evidence, "double")
-  expect_named(res, c("log_evidence", "x", "y", "z"))
-  expect_snapshot(res)
+  expect_named(res, c("id", "split", "run"))
+  expect_type(res$id, "integer")
+  expect_type(res$split, "list")
+  expect_type(res$run, "list")
 })
 
-test_that("learn is reproducible with a fixed seed", {
+test_that("generate_bootstraps is reproducible with a fixed seed", {
   withr::local_seed(123)
-  out1 <- learn(example_run, times = 5)
+  out1 <- generate_bootstraps(example_run, times = 5)
 
   withr::local_seed(123)
-  out2 <- learn(example_run, times = 5)
+  out2 <- generate_bootstraps(example_run, times = 5)
 
   expect_identical(out1, out2)
 })
 
-test_that("learn returns include_weights", {
-  res <- learn(example_run, times = 4, include_weights = TRUE)
-  expect_equal(nrow(res), 4)
-  expect_type(res$weights, "list")
-  expect_all_equal(vapply(res$weights, \(x) sum(x$weight), double(1)), 1)
+test_that("generate_bootstraps handles times = 0", {
+  res0 <- generate_bootstraps(example_run, times = 0)
+  expect_equal(nrow(res0), 0)
 })
 
 #' @srrstats {G5.0, G5.6, G5.6a, BS7.2} Use the NIST dataset `mcmc01` to test
@@ -110,14 +103,21 @@ describe("Parameter recovery", {
     expect_lt(median_sd, sd_upper)
   })
 
-  it("Recovers parameters in the ernest_resample object", {
-    bs <- learn(run)
-
-    median_mean <- median(bs$mu)
+  it("Recovers parameters via bootstrapping", {
+    bs <- generate_bootstraps(run, times = 100)
+    means <- vapply(
+      bs$run,
+      \(run) {
+        m <- posterior::summarise_draws(run, "mean")$mean
+        names(m) <- c("mu", "sigma")
+        m
+      },
+      numeric(2)
+    )
+    median_mean <- median(means["mu", ])
     expect_gt(median_mean, mean_lower)
     expect_lt(median_mean, mean_upper)
-
-    median_sd <- median(bs$sigma)
+    median_sd <- median(means["sigma", ])
     expect_gt(median_sd, sd_lower)
     expect_lt(median_sd, sd_upper)
   })
