+++ R/boot.R
@@ -1,0 +1,230 @@
+#' Confidence intervals for model evidence
+#'
+#' @inheritParams generate_bootstraps
+#' @param parm `[character(1)]`\cr A specification describing whether to return
+#' the model evidence in log-units.
+#' @param level `[double(1)]`\cr The confidence level required.
+#' @inheritParams rlang::check_dots_empty
+#'
+#' @returns A named vector with two elements, describing the lower and upper
+#' confidence limits for each parameter.
+#'
+#' @seealso [generate_bootstraps()] for generating resampled nested sampling
+#' runs from an `ernest_run` object.
+#' @export
+confint.ernest_run <- function(
+  object,
+  parm = c("log_evidence", "evidence"),
+  level = 0.95,
+  times = 100,
+  ...
+) {
+  check_dots_empty()
+  parm <- arg_match(parm)
+  bootstraps <- generate_bootstraps(object, times = times)
+  log_evidence <- vapply(bootstraps$run, log_evidence, double(1))
+  if (parm == "evidence") {
+    log_evidence <- exp(log_evidence)
+  }
+  stats::quantile(log_evidence, probs = c((1 - level) / 2, (1 + level) / 2))
+}
+
+#' Bootstrap resampling over a nested sampling run.
+#'
+#' @param object [[ernest_run]]\cr An object containing the results of a nested
+#' sampling run.
+#' @param times `[[integer(1)]]`\cr The number of bootstrap resamples.
+#' @param units `[character(1)]`\cr The scale of the sampled points reported in
+#' the resample; see [as_draws.ernest].
+#' @param apparent `[[logical(1)]]`\cr Whether to perform an additional resample
+#' using a copy of the original run data stored in `x`.
+#' @inheritDotParams rlang::check_dots_empty
+#'
+#' @returns A tibble. The results include a column for an [ernest_rcrd] object
+#' containing the resampled run, and a column called `id` that has a character
+#' string with the resample identifier.
+#'
+#' @details
+#' Higson et al. (2019) describes a bootstrap resampling procedure for nested
+#' sampling. This involves splitting a run into `nlive` runs of one live point,
+#' then merging runs together by sampling from these runs with replacement.
+#' This provides an empirical estimate of the uncertainty due to the
+#' stochastic nature of selecting likelihood shells from the parameter space.
+#'
+#' @seealso [calculate.ernest_run()] for evidence estimation without
+#' bootstrapping.
+#'
+#' @references
+#' * Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
+#' Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
+#' of the Royal Astronomical Society, 483(2), 2044–2056.
+#' \doi{10.1093/mnras/sty3090}
+#' * Speagle, J. S. (2020). dynesty: A Dynamic Nested Sampling Package for
+#' Estimating Bayesian Posteriors and Evidences. Monthly Notices of the Royal
+#' Astronomical Society, 493, 3132–3158. \doi{10.1093/mnras/staa278}
+#'
+#' @srrstats {BS4.2} Allows for validating posterior estimates through
+#' bootstrapping the importance weights.
+#'
+#' @examples
+#' # Generate 100 bootstrap resamples
+#' generate_bootstraps(example_run)
+#' @export
+generate_bootstraps <- function(
+  x,
+  times = 100,
+  units = c("original", "unit_cube"),
[…]
