### 2026-05-05 8c76eac
New bootstrap method, helpers for accessing common features from ernest_rcrd and ernest_run.
