+++ R/learn.R
@@ -1,175 +1,0 @@
-#' Bootstrap Resampling for Nested Sampling Uncertainty
-#'
-#' Computes bootstrap resamples of a nested sampling run to empirically
-#' estimate uncertainty in evidence and posterior summaries.
-#'
-#' @param x [[ernest_run]]\cr An object containing the results of a nested
-#' sampling run.
-#' @param times `[[integer(1)]]`\cr Number of bootstrap resamples to generate.
-#' Must be larger than zero.
-#' @param include_weights `[[logical(1)]]`\cr If `TRUE`, include a list-column
-#' of resampled weights for each resampled run in the output.
-#' @inheritParams as_draws.ernest_run
-#' @inheritParams rlang::args_dots_empty
-#'
-#' @returns An `ernest_resample` object, which inherits from `tbl_df`, `tbl`,
-#' and `data.frame`.
-#'
-#' Each row corresponds to a bootstrap resample of the nested with one
-#' row per resample, containing:
-#' * `niter`: Number of iterations in the resample.
-#' * `log_evidence`: Estimated log-evidence for the resample.
-#' * One column per parameter: weighted posterior mean for each parameter.
-#' * If `include_weights = TRUE`, a list-column `weights` with the resampled
-#' weights and log-volumes for each resample.
-#'
-#' @details
-#' Higson et al. (2019) describes a bootstrap resampling procedure for nested
-#' sampling. This involves splitting a run into `nlive` runs of one live point,
-#' then merging runs together by sampling from these runs with replacement.
-#' This provides an empirical estimate of the uncertainty due to the
-#' stochastic nature of selecting likelihood shells from the parameter space.
-#'
-#' @seealso [calculate.ernest_run()] for evidence estimation without
-#' bootstrapping.
-#'
-#' @references
-#' * Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
-#' Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
-#' of the Royal Astronomical Society, 483(2), 2044–2056.
-#' \doi{10.1093/mnras/sty3090}
-#' * Speagle, J. S. (2020). dynesty: A Dynamic Nested Sampling Package for
-#' Estimating Bayesian Posteriors and Evidences. Monthly Notices of the Royal
-#' Astronomical Society, 493, 3132–3158. \doi{10.1093/mnras/staa278}
-#'
-#' @srrstats {BS4.2} Allows for validating posterior estimates through
-#' bootstrapping the importance weights.
-#'
-#' @examples
-#' data(example_run)
-#' # Generate 100 bootstrap resamples
-#' learn(example_run, times = 100)
-#' # Include resampled weights for each draw
-#' learn(example_run, times = 10, include_weights = TRUE)
-#' @export
-learn.ernest_run <- function(
-  x,
-  times = 100,
-  include_weights = FALSE,
-  units = c("original", "unit_cube"),
-  ...
-) {
-  check_dots_empty()
-  check_number_whole(times, min = 1)
-  check_bool(include_weights)
-  units <- arg_match(units)
-  x_rcrd <- x$rcrd
-  dead_log_vol <- get_log_vol(x_rcrd)[x$niter]
-
-  field(x_rcrd, "unit") <- as_draws_matrix_(
-    x,
-    units = units,
-    radial = FALSE
-  )$points
-  threads <- get_threads(x_rcrd)
-  res <- replicate(
-    times,
-    run_resample(x$nlive, threads, x$prior$names, include_weights),
-    simplify = FALSE
-  )
-
-  new_tibble0(
-    x = vctrs::vec_rbind(!!!res),
[…]
