+++ R/boot.R
@@ -1,0 +1,206 @@
+#' Bootstrap resampling over a nested sampling run.
+#'
+#' @param object [[ernest_run]]\cr An object containing the results of a nested
+#' sampling run.
+#' @param times `[[integer(1)]]`\cr The number of bootstrap resamples.
+#' @param units `[character(1)]`\cr The scale of the sampled points reported in
+#' the resample; see [as_draws.ernest].
+#' @param apparent `[[logical(1)]]`\cr Whether to perform an additional resample
+#' using a copy of the original run data stored in `x`.
+#' @inheritDotParams rlang::check_dots_empty
+#'
+#' @returns A tibble. The results include a column for an [ernest_rcrd] object
+#' containing the resampled run, and a column called `id` that has a character
+#' string with the resample identifier.
+#'
+#' @details
+#' Higson et al. (2019) describes a bootstrap resampling procedure for nested
+#' sampling. This involves splitting a run into `nlive` runs of one live point,
+#' then merging runs together by sampling from these runs with replacement.
+#' This provides an empirical estimate of the uncertainty due to the
+#' stochastic nature of selecting likelihood shells from the parameter space.
+#'
+#' @seealso [calculate.ernest_run()] for evidence estimation without
+#' bootstrapping.
+#'
+#' @references
+#' * Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
+#' Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
+#' of the Royal Astronomical Society, 483(2), 2044–2056.
+#' \doi{10.1093/mnras/sty3090}
+#' * Speagle, J. S. (2020). dynesty: A Dynamic Nested Sampling Package for
+#' Estimating Bayesian Posteriors and Evidences. Monthly Notices of the Royal
+#' Astronomical Society, 493, 3132–3158. \doi{10.1093/mnras/staa278}
+#'
+#' @srrstats {BS4.2} Allows for validating posterior estimates through
+#' bootstrapping the importance weights.
+#'
+#' @examples
+#' # Generate 100 bootstrap resamples
+#' generate_bootstraps(example_run)
+#' @export
+generate_bootstraps <- function(
+  x,
+  times = 100,
+  units = c("original", "unit_cube"),
+  apparent = FALSE,
+  ...
+) {
+  check_dots_empty()
+  check_number_whole(times, min = 0)
+  units <- arg_match(units)
+
+  x_rcrd <- x$rcrd
+  threads <- get_threads(x_rcrd)
+  nthreads <- vctrs::vec_size(threads)
+  sample_ids <- replicate(
+    times,
+    sample.int(nthreads, replace = TRUE),
+    simplify = FALSE
+  )
+
+  field(x_rcrd, "unit") <- as_draws_matrix_(
+    x,
+    units = units,
+    radial = FALSE
+  )$points
+  resampled <- resample_runs(x_rcrd, threads, x$nlive, sample_ids)
+  sample_ids <- lapply(
+    sample_ids,
+    \(x) vctrs::vec_unrep(sort.int(x))
+  )
+  id <- paste0("Bootstrap", seq_along(sample_ids))
+
+  if (apparent) {
+    id <- c(id, "Apparent")
+    sample_ids <- c(sample_ids, list(seq(x$nlive)))
+    resampled <- c(resampled, list(x$rcrd))
+  }
+  new_tibble0(
+    data_frame0(
+      "id" = id,
+      "split" = sample_ids,
+      "run" = resampled
+    )
+  )
+}
+
+#' Split run into live point threads
+#'
[…]

[…]
