### 2026-05-08 736d6bc
Rollback bootstrapping, merging, and parallelization.
