+++ R/merge.R
@@ -1,164 +1,0 @@
-#' Merge two nested sampling runs
-#'
-#' Combines two nested sampling runs into a single run by re-ordering their
-#' samples by log-likelihood and reconstructing the live set.
-#'
-#' @param x,y [[ernest_run]]\cr Runs generated from compatible nested sampling
-#' specifications.
-#' @inheritParams rlang::args_dots_empty
-#'
-#' @returns [[ernest_run]] A run containing the merged results from `x` and `y`.
-#'
-#' @details
-#' The two runs must use the same prior variable names and LRPS method.
-#' If `first_update` or `update_interval` differ between runs, the defaults
-#' from [[ernest_sampler]] are used instead.
-#'
-#' The merged run contains all dead points from both runs, ordered by
-#' log-likelihood, and the live set is reconstructed from the remaining points.
-#' This results in a single nested sampling run with `nlive` equal to the sum of
-#' the two original runs.
-#'
-#' Note that `niter` is not simply the sum of the two original runs' `niter`
-#' values; rather, the merged run's `niter` is equal to the total number of
-#' samples drawn before either run's termination condition was met. This means
-#' that the merged run's `niter` reflects the number of iterations it would take
-#' to reach the same likelihood contour as the least advanced of the two
-#' original runs.
-#'
-#' If the two runs do not share the same RNG seed, the merged result stores
-#' `NA` as the seed.
-#'
-#' @examples
-#' prior <- create_uniform_prior(lower = c(-1, -1), upper = 1)
-#' ll_fn <- function(x) -sum(x^2)
-#'
-#' run1 <- generate(
-#'   ernest_sampler(ll_fn, prior, nlive = 100, seed = 24),
-#'   max_iterations = 100
-#' )
-#' run2 <- generate(
-#'   ernest_sampler(ll_fn, prior, nlive = 300, seed = 42),
-#'   max_iterations = 100
-#' )
-#' merged <- merge(run1, run2)
-#' summary(merged)
-#' @export
-merge.ernest_run <- function(x, y, ...) {
-  xy <- reindex_runs(x$rcrd, y$rcrd)
-  results <- merge_results(!!!xy)
-  z <- merge_sampler(x, y)
-  # Reform sampler
-  env_bind(
-    z$live_env,
-    !!!as.list(results$live)[c("unit", "log_lik", "birth_lik")]
-  )
-  z <- refresh_ernest_sampler(z)
-  new_ernest_run_(z, results$dead)
-}
-
-#' Reindex runs so they have continuous IDs from 1 to 'nlive'
-#'
-#' @param ... ernest_rcrd objects.
-#'
-#' @return A list of ernest_rcrd objects with reindexed IDs, same length as
-#' the dots.
-#' @noRd
-reindex_runs <- function(...) {
-  runs <- list2(...)
-  nlives <- vapply(
-    runs,
-    \(x) vctrs::vec_unique_count(field(x, "id")),
-    integer(1)
-  )
-  .mapply(
-    \(x, start) {
-      field(x, "id") <- as.integer(
-        vctrs::vec_group_id(field(x, "id")) + start
-      )
-      x
-    },
-    dots = list(runs, c(0, cumsum(nlives)[-length(nlives)])),
-    MoreArgs = NULL
-  )
-}
-
-#' Merge the dead points from multiple nested sampling runs.
-#'
-#' @param ... A set of ernest_rcrd. Each rcrd should have unique
-#' IDs, else an error is thrown.
-#'
-#' @importFrom vctrs `field<-`
[…]
