+++ tests/testthat/_snaps/learn.md
@@ -1,20 +1,0 @@
-# learn returns expected structure
-
-    Code
-      res
-    Output
-      # A tibble: 100 x 4
-         log_evidence        x         y         z
-                <dbl>    <dbl>     <dbl>     <dbl>
-       1        -8.95 -0.0177  -0.000244 -0.00863 
-       2        -9.01 -0.0175  -0.0151    0.00297 
-       3        -8.99 -0.0399   0.00547   0.00715 
-       4        -8.97 -0.0312  -0.0206    0.00720 
-       5        -8.99 -0.0405  -0.0317   -0.00486 
-       6        -9.11 -0.0296   0.0160    0.0120  
-       7        -9.11 -0.00761  0.00744   0.000965
-       8        -8.92 -0.0353  -0.0132    0.0152  
-       9        -9.05 -0.00489 -0.0144    0.0144  
-      10        -9.03 -0.0204  -0.0476   -0.000800
-      # i 90 more rows
-

[…]
