### 2026-05-08 86a757b
Rename `n_dim` to `nvar` for clarity/conformity with other named elements of `ernest_run`.
