+++ tests/testthat/test-unif_ellipsoid.R
@@ -20,7 +20,7 @@ describe("unif_ellipsoid class", {
     obj <- expect_all_proposals(
       new_unif_ellipsoid,
       unit_log_fn = fn,
-      n_dim = 2,
+      nvar = 2,
       allow_failure = TRUE
     )
     expect_lrps(obj, subclass = "unif_ellipsoid", !!!ptypes)
@@ -31,7 +31,7 @@ describe("unif_ellipsoid class", {
   })
 
   it("Can be updated with a matrix of points", {
-    obj <- new_unif_ellipsoid(fn, n_dim = 2)
+    obj <- new_unif_ellipsoid(fn, nvar = 2)
     samples <- run_sampler(obj)
     expect_lrps(obj, subclass = "unif_ellipsoid", !!!ptypes)
     original_volume <- obj$cache$log_volume
@@ -47,7 +47,7 @@ describe("unif_ellipsoid class", {
   })
 
   it("Can be updated without a matrix", {
-    obj <- new_unif_ellipsoid(fn, n_dim = 2)
+    obj <- new_unif_ellipsoid(fn, nvar = 2)
     samples <- run_sampler(obj)
     expect_idempotent_update(
       obj,
@@ -57,7 +57,7 @@ describe("unif_ellipsoid class", {
   })
 
   it("Throws a warning when provided with constant points", {
-    obj <- new_unif_ellipsoid(fn, n_dim = 2)
+    obj <- new_unif_ellipsoid(fn, nvar = 2)
     live <- matrix(rep(0.5, 500 * 2), nrow = 500)
     expect_warning(
       update_lrps(obj, live),
