### 2026-05-08 16a571d
Rename 'rcrd' element; streamline as_draws method.
