+++ tests/testthat/test-visualize.R
@@ -35,11 +35,6 @@ test_that("Density plot", {
     "visualize-yz",
     visualize(example_run, y, z, .units = "unit_cube")
   )
-
-  vdiffr::expect_doppelganger(
-    "visualize-radial",
-    visualize(example_run, ".radial", .radial = TRUE)
-  )
 })
 
 test_that("Trace plot", {
@@ -57,10 +52,5 @@ test_that("Trace plot", {
   vdiffr::expect_doppelganger(
     "trace-yz",
     visualize(example_run, y, z, .which = "trace", .units = "unit_cube")
-  )
-
-  vdiffr::expect_doppelganger(
-    "trace-radial",
-    visualize(example_run, .which = "trace", .radial = TRUE)
   )
 })
