### 2026-05-08 5dd6f79
Tests and documentation for the weights function.
