+++ R/as_draws.R
@@ -87,58 +87,3 @@ as_draws_matrix.ernest_run <- function(
     log = TRUE
   )
 }
-
-#' Extract the posterior importance weights from a nested sampling run
-#'
-#' Return the normalised importance weights for the dead points in a nested
-#' sampling run. On the log scale, these are `log_weight - log_evidence`; on
-#' the probability scale, they are exponentiated so they sum to one.
-#'
-#' @param x [[ernest_run]]\cr A nested sampling run.
-#' @param log `[logical(1)]`\cr Whether the weights should be returned on the
-#' log scale.
-#' @inheritParams rlang::args_dots_empty
-#'
-#' @returns `[double()]` A numeric vector of normalised importance weights.
-#'
-#' @details
-#' The log-weights in a nested sampling run are the individual contributions
-#' of each sample to the log-evidence estimate. They are calculated from the
-#' sample's log-likelihood and the amount of prior volume inside its likelihood
-#' contour.
-#'
-#' The posterior importance weights are obtained by normalising the log-weights
-#' with the final log-evidence estimate. They can be used to reweight the
-#' posterior samples from the run so they approximate the posterior
-#' distribution.
-#'
-#' @seealso [as_draws.ernest_run] to extract the posterior samples from a nested
-#' sampling run, bound to their corresponding importance weights.
-#'
-#' @exportS3Method stats::weights
-#' @examples
-#' data(example_run)
-#' weights(example_run) |> head()
-#' weights(example_run, log = TRUE) |> head()
-#' @export
-weights.ernest_run <- function(x, log = FALSE, ...) {
-  check_dots_empty()
-  weights <- x$log_weight - x$log_evidence
-  if (log) {
-    weights
-  } else {
-    exp(weights)
-  }
-}
-
-#' @noRd
-#' @export
-weights.ernest_rcrd <- function(x, log = FALSE, ...) {
-  log_w <- compute_integral(x, truncate = TRUE)
-  weights <- log_w$log_weight - log_w$log_evidence
-  if (log) {
-    weights
-  } else {
-    exp(weights)
-  }
-}
