+++ man/weights.ernest_run.Rd
@@ -1,37 +1,38 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/as_draws.R
+% Please edit documentation in R/weights.R
 \name{weights.ernest_run}
 \alias{weights.ernest_run}
-\title{Extract the posterior importance weights from a nested sampling run}
+\title{Extract the posterior sample weights from a nested sampling run}
 \usage{
 \method{weights}{ernest_run}(x, log = FALSE, ...)
 }
 \arguments{
 \item{x}{[\link{ernest_run}]\cr A nested sampling run.}
 
-\item{log}{\verb{[logical(1)]}\cr Whether the weights should be returned on the
-log scale.}
+\item{log}{\verb{[logical(1)]}\cr Whether to return the weights on the log
+scale.}
 
 \item{...}{These dots are for future extensions and must be empty.}
 }
 \value{
-\verb{[double()]} A numeric vector of normalised importance weights.
+\verb{[double()]} A numeric vector of normalised importance weights. When
+\code{log = FALSE}, the values are exponentiated so they sum to one.
 }
 \description{
-Return the normalised importance weights for the dead points in a nested
-sampling run. On the log scale, these are \code{log_weight - log_evidence}; on
-the probability scale, they are exponentiated so they sum to one.
+Return the normalised posterior importance weights for the dead points in a
+nested sampling run.
 }
 \details{
 The log-weights in a nested sampling run are the individual contributions
-of each sample to the log-evidence estimate. They are calculated from the
-sample's log-likelihood and the amount of prior volume inside its likelihood
-contour.
+of each sample to the log-evidence estimate. The unnormalised weight of the
+\eqn{i}th sampled point is given as
+\deqn{w_i = \frac{L_{i-1} + L_i}{2} * (V_{i-1} - V_i)}
+where \eqn{L_i} is the likelihood value for the point and \eqn{V_i} is the
+prior volume at which the point was sampled.
 
 The posterior importance weights are obtained by normalising the log-weights
-with the final log-evidence estimate. They can be used to reweight the
-posterior samples from the run so they approximate the posterior
-distribution.
+with the final log-evidence estimate. They can be used to reweight posterior
+samples from the run so they approximate the posterior distribution.
 }
 \examples{
 data(example_run)

[…]
