+++ R/likelihood-class.R
@@ -45,9 +45,8 @@
 #' (see `on_nonfinite`). The same parameter also controls behaviour for
 #' non-finite, non-`-Inf` values.
 #'
-#' @seealso See the
-#' [cubature](https://bnaras.github.io/cubature/articles/cubature.html) package
-#' for more examples of scalar and vectorized functions.
+#' @seealso [cubature](https://bnaras.github.io/cubature/articles/cubature.html)
+#' for examples of vectorized and non-vectorized likelihood functions.
 #'
 #' @aliases ernest_likelihood
 #' @example ./data-raw/EXAMPLE_LIKELIHOOD.R
