+++ R/plot.R
@@ -46,9 +46,8 @@
 #'
 #' @srrstats {BS6.1} Default plot for return object.
 #'
-#' @seealso
-#' * [calculate()] for generating `ernest_estimate` objects.
-#' * [visualize()] for plotting posterior distributions generated by a run.
+#' @seealso [calculate.ernest_run()]
+#' @family visualizations
 #'
 #' @examples
 #' # Plot diagnostics from a run (analytical uncertainty for evidence).
@@ -59,7 +58,7 @@
 #' set.seed(123)
 #' est <- calculate(example_run, ndraws = 100)
 #' plot(est)
-#' @rdname plot.ernest
+#' @rdname plot-ernest
 #' @export
 plot.ernest_estimate <- function(
   x,
@@ -71,7 +70,7 @@ plot.ernest_estimate <- function(
   print(autoplot(x, which, n = n, ...))
 }
 
-#' @rdname plot.ernest
+#' @rdname plot-ernest
 #' @export
 plot.ernest_run <- function(
   x,
@@ -88,7 +87,7 @@ plot.ernest_run <- function(
 
 #' @rdname plot.ernest
 #'
-#' @param object [[ernest_estimate]]\cr Output from [calculate()].
+#' @param object [[ernest_estimate]]\cr Output from [calculate.ernest_run()].
 #' @param which `[character()]`\cr One or more diagnostics to summarize. Must be
 #' any of `"evidence"`, `"weight"`, and `"likelihood"`.
 #' @param width `[double()]`\cr Vector of probabilities to use that determine
