+++ R/import-generics.R
@@ -13,7 +13,3 @@ generics::generate
 #' @importFrom generics visualize
 #' @export
 generics::visualize
-
-#' @importFrom generics learn
-#' @export
-generics::learn
