+++ R/ernest-package.R
@@ -10,6 +10,3 @@
 #' @useDynLib ernest, .registration = TRUE
 ## usethis namespace: end
 NULL
-
-# CRAN flags on parallelized variables
-utils::globalVariables(c("impl_", "lrps_"))
