### 2026-05-11 2013af4
Spellcheck, updated codemeta, added ROpenSci reviewers to DESCRIPTION.
