+++ DESCRIPTION
@@ -1,9 +1,12 @@
 Package: ernest
 Title: A Toolkit for Nested Sampling
 Version: 1.2.0
-Authors@R: 
+Authors@R: c(
     person("Kyle", "Dewsnap", , "kyle.dewsnap@ubc.ca", role = c("aut", "cre"),
-           comment = c(ORCID = "0000-0003-2132-8083"))
+           comment = c(ORCID = "0000-0003-2132-8083")),
+    person("TJ", "Mahr", role = "ctb"),
+    person("Robert", "Kubinec", role = "ctb")
+  )
 Description: Bayesian evidence estimation and posterior inference with the
     nested sampling algorithm, along with S3 methods for simulating
     uncertainty and creating visualisations.

[…]
