### 2026-05-11 9d387f3
Fix vignette bug when printing long rvar tibbles.
