+++ vignettes/more-ernest-runs.Rmd
@@ -66,7 +66,8 @@ The expected log-evidence, found analytically, is $-6.679$. We may use the summa
 
 ```{r}
 summary(blob_result)
-calculate(blob_result, ndraws = 500)
+calculate(blob_result, ndraws = 500)$log_evidence |>
+  tail(1)
 ```
 
 ```{r include = FALSE}
