+++ R/as_draws.R
@@ -30,6 +30,7 @@
 #'
 #' # View importance weights
 #' dm <- as_draws(example_run)
+#' str(dm)
 #' weights(dm) |> head()
 #'
 #' # Summarise points after resampling
@@ -37,13 +38,9 @@
 #'   resample_draws() |>
 #'   summarize_draws()
 #'
-#' # View the radial coordinate in unit space over the run
-#' dm_rad <- as_draws_rvars(
-#'   example_run,
-#'   units = "unit_cube",
-#'   radial = TRUE
-#' )
-#' plot(x = draws_of(dm_rad$.radial))
+#' # Extract the same coordinates in the unit space coordinates
+#' dm_unit <- as_draws_rvars(example_run, units = "unit_cube")
+#' str(dm_unit)
 #' @export
 as_draws.ernest_run <- function(
   x,
