+++ reference/generate-ernest.md
@@ -1,0 +1,166 @@
+# Run nested sampling to estimate Bayesian evidence
+
+Executes the nested sampling algorithm, iteratively replacing the worst
+live point with a new sample from a likelihood-restricted prior until a
+stopping criterion is met.
+
+## Usage
+
+``` r
+# S3 method for class 'ernest_sampler'
+generate(
+  x,
+  max_iterations = NULL,
+  max_evaluations = NULL,
+  min_logz = 0.05,
+  show_progress = NULL,
+  ...
+)
+
+# S3 method for class 'ernest_run'
+generate(
+  x,
+  max_iterations = NULL,
+  max_evaluations = NULL,
+  min_logz = 0.05,
+  show_progress = NULL,
+  ...
+)
+```
+
+## Arguments
+
+- x:
+
+  \[[ernest_sampler](https://kylesnap.github.io/ernest/reference/ernest_sampler.md)\]
+  or \[ernest_run\]  
+  A nested sampling specification.
+
+- max_iterations:
+
+  `[integer(1)]`  
+  The maximum number of iterations to perform. Optional; if `NULL` this
+  criterion is ignored.
+
+- max_evaluations:
+
+  `[integer(1)]`  
+  The maximum number of times the run can evaluate the likelihood
+  function. Optional; if `NULL` this criterion is ignored.
+
+- min_logz:
+
+  `[double(1)]`  
+  The minimum log-ratio between the current estimated evidence and the
+  remaining evidence. Must be non-negative; if set to zero, this
+  criterion is ignored.
+
+- show_progress:
+
+  `[logical(1)]`  
+  If `TRUE`, displays a progress spinner and iteration counter during
+  sampling. Optional; if `NULL` the global option
+  `rlib_message_verbosity` is used to determine whether to show
+  progress.
+
+- ...:
+
+  Arguments passed on to
+  [`compile.ernest_run`](https://kylesnap.github.io/ernest/reference/compile-ernest.md)
+
+  `clear`
+
+  :   `[logical(1)]`  
+      If `TRUE`, clears results from previous runs before compiling. If
+      `FALSE`, retains previous results and validates the live set.
+
+## Value
+
+An `[ernest_run]` object with the nested sampling results.
+
+This inherits from
+[ernest_sampler](https://kylesnap.github.io/ernest/reference/ernest_sampler.md)
+and adds:
+
+- `niter`: `[integer(1)]` Number of iterations performed.
+
+- `neval`: `[integer(1)]` Number of times the likelihood function was
+  evaluated.
+
+- `log_evidence`: `[double(1)]` The log-evidence estimate.
+
+- `log_evidence_err`: `[double(1)]` The standard error of the
+  log-evidence estimate, derived using `information`.
+
+- `information`: `[double(1)]` The estimated Kullback-Leibler
+  divergence.
+
+- `log_weight`: `[double(nsample)]` Each sample's contribution to the
+  log-evidence estimate, computed from its log-likelihood and prior
+  volume.
+
+- `rcrd`:
+  \[[ernest_rcrd](https://kylesnap.github.io/ernest/reference/ernest_rcrd.md)\]
+  An object storing an internal record of each point generated during
+  the run.
+
+## Details
+
+At least one of `max_iterations`, `max_evaluations`, or `min_logz` must
+specify a valid stopping criterion. Setting `min_logz` to zero while
+leaving `max_iterations` and `max_evaluations` at their defaults will
+result in an error.
+
+If `x` is an `ernest_run` object, the stopping criteria are checked
+against the current state of the run. An error is thrown if the stopping
+criteria have already been satisfied by `x`.
+
[…]
