### 2026-05-11 e2bc1dc
Update pkgcheck workflow triggers
Changed trigger from push and pull_request to workflow_run.
