### 2026-05-13 01bd5bc
"id" no longer re-sorted at compile or at disassembly of live_env.
