+++ tests/testthat/_snaps/run-class.md
@@ -29,15 +29,15 @@
       -- Posterior Summary -----------------------------------------------------------
     Output
       # A tibble: # x #
-        variable     mean    sd   median   q#   q#
-        <chr>       <dbl> <dbl>    <dbl> <dbl> <dbl>
-      # x        -#.#   #.#  #.# -#.#  #.#
-      # y        -#.#  #.# -#.# -#.#  #.#
-      # z         #.#   #.#  #.# -#.#  #.#
+        variable     mean    sd    median   q#   q#
+        <chr>       <dbl> <dbl>     <dbl> <dbl> <dbl>
+      # x         #.#  #.# -#.#  -#.#  #.#
+      # y        -#.#  #.#  #.# -#.#  #.#
+      # z        -#.#   #.# -#.#   -#.#  #.#
     Message
       -- Maximum Likelihood Estimate (MLE) -------------------------------------------
       * Log-likelihood: -#.#
-      * Original parameters: #.#, -#.#, and -#.#
+      * Original parameters: -#.#, #.#, and -#.#
 
 # summary.ernest_run returns expected structure and values / has the expected posterior
 
