+++ tests/testthat/_snaps/plot.md
@@ -83,7 +83,7 @@
         {
           "type": "double",
           "attributes": {},
-          "value": [-16.84447086, -16.81150712, -16.77854338, -16.74557964, -16.7126159, -16.67965216, -16.64668842, -16.61372468, -16.58076094, -16.54779721, -16.51483347, -16.48186973, -16.44890599, -16.41594225, -16.38297851, -16.35001477, -16.31705103, -16.28408729, -16.25112355, -16.21815981, -16.18519607, -16.15223233, -16.11926859, -16.08630485, -16.05334111, -16.02037737, -15.98741363, -15.9544499, -15.92148616, -15.88852242, -15.85555868, -15.82259494, -15.7896312, -15.75666746, -15.72370372, -15.69073998, -15.65777624, -15.6248125, -15.59184876, -15.55888502, -15.52592128, -15.49295754, -15.4599938, -15.42703006, -15.39406632, -15.36110259, -15.32813885, -15.29517511, -15.26221137, -15.22924763, -15.19628389, -15.16332015, -15.13035641, -15.09739267, -15.06442893, -15.03146519, -14.99850145, -14.96553771, -14.93257397, -14.89961023, -14.86664649, -14.83368275, -14.80071901, -14.76775528, -14.73479154, -14.7018278, -14.66886406, -14.63590032, -14.60293658, -14.56997284, -14.5370091, -14.50404536, -14.47108162, -14.43811788, -14.40515414, -14.3721904, -14.33922666, -14.30626292, -14.27329918, -14.24033544, -14.2073717, -14.17440796, -14.14144423, -14.10848049, -14.07551675, -14.04255301, -14.00958927, -13.97662553, -13.94366179, -13.91069805, -13.87773431, -13.84477057, -13.81180683, -13.77884309, -13.74587935, -13.71291561, -13.67995187, -13.64698813, -13.61402439, -13.58106065, -13.54809692, -13.51513318, -13.48216944, -13.4492057, -13.41624196, -13.38327822, -13.35031448, -13.31735074, -13.284387, -13.25142326, -13.21845952, -13.18549578, -13.15253204, -13.1195683, -13.08660456, -13.05364082, -13.02067708, -12.98771334, -12.95474961, -12.92178587, -12.88882213, -12.85585839, -12.82289465, -12.78993091, -12.75696717, -12.72400343, -12.69103969, -12.65807595, -12.62511221, -12.59214847, -12.55918473, -12.52622099, -12.49325725, -12.46029351, -12.42732977, -12.39436603, -12.36140229, -12.32843856, -12.29547482, -12.26251108, -12.22954734, -12.1965836, -12.16361986, -12.13065612, -12.09769238, -12.06472864, -12.0317649, -11.99880116, -11.96583742, -11.93287368, -11.89990994, -11.8669462, -11.83398246, -11.80101872, -11.76805498, -11.73509125, -11.70212751, -11.66916377, -11.63620003, -11.60323629, -11.57027255, -11.53730881, -11.50434507, -11.47138133, -11.43841759, -11.40545385, -11.37249011, -11.33952637, -11.30656263, -11.27359889, -11.24063515, -11.20767141, -11.17470767, -11.14174394, -11.1087802, -11.07581646, -11.04285272, -11.00988898, -10.97692524, -10.9439615, -10.91099776, -10.87803402, -10.84507028, -10.81210654, -10.7791428, -10.74617906, -10.71321532, -10.68025158, -10.64728784, -10.6143241, -10.58136036, -10.54839663, -10.51
[…]
