+++ articles/nested-sampling-with-ernest.md
@@ -439,8 +439,8 @@ run_1k
 #> ── Results ─────────────────────────────────────────────────────────────────────
 #> * Iterations: 1000
 #> * Likelihood evals.: 16401
-#> * Log-evidence: -983.8285 (± 23.0942)
-#> * Information: 533.3
+#> * Log-evidence: -1614.9184 (± 27.7466)
+#> * Information: 802.4
 ```
 
 `generate` produces an `ernest_run` object, which inherits from
@@ -554,10 +554,10 @@ run
 #> * Prior: normal prior distribution with 5 dimensions (Intercept, zAge, zBase,
 #> Trt1, and zBase:Trt1)
 #> ── Results ─────────────────────────────────────────────────────────────────────
-#> * Iterations: 7977
-#> * Likelihood evals.: 190826
-#> * Log-evidence: -883.6069 (± 0.3218)
-#> * Information: 21.31
+#> * Iterations: 7872
+#> * Likelihood evals.: 188201
+#> * Log-evidence: -883.1998 (± 0.3181)
+#> * Information: 20.86
 ```
 
 ### Uncertainty
@@ -595,23 +595,23 @@ summary(run)
 #> Summary of nested sampling run:
 #> ── Run Information ─────────────────────────────────────────────────────────────
 #> * No. points: 300
-#> * Iterations: 7977
-#> * Likelihood evals.: 190826
-#> * Log-evidence: -883.6069 (± 0.3218)
-#> * Information: 21.31
+#> * Iterations: 7872
+#> * Likelihood evals.: 188201
+#> * Log-evidence: -883.1998 (± 0.3181)
+#> * Information: 20.86
 #> * RNG seed: 42
 #> ── Posterior Summary ───────────────────────────────────────────────────────────
 #> # A tibble: 5 × 6
 #>   variable      mean    sd  median    q15   q85
 #>   <chr>        <dbl> <dbl>   <dbl>  <dbl> <dbl>
-#> 1 Intercept   1.39   1.23   1.89    0.539 2.01 
-#> 2 zAge        0.101  0.832  0.149  -0.178 0.399
-#> 3 zBase       0.445  0.842  0.570   0.207 0.779
-#> 4 Trt1       -0.157  1.06  -0.180  -0.647 0.406
-#> 5 zBase:Trt1  0.0153 0.870  0.0463 -0.293 0.343
+#> 1 Intercept   1.41   1.22   1.89    0.623 2.01 
+#> 2 zAge        0.112  0.841  0.151  -0.155 0.416
+#> 3 zBase       0.445  0.832  0.567   0.195 0.768
+#> 4 Trt1       -0.184  1.05  -0.185  -0.641 0.363
+#> 5 zBase:Trt1  0.0254 0.856  0.0480 -0.263 0.369
 #> ── Maximum Likelihood Estimate (MLE) ───────────────────────────────────────────
-#> * Log-likelihood: -860.034
-#> * Original parameters: 1.9441, 0.1469, 0.5751, -0.2006, and 0.0442
+#> * Log-likelihood: -859.9751
[…]

[…]
