+++ R/merge.R
@@ -1,0 +1,145 @@
+#' Merge `ernest_run` objects together.
+#'
+#' Combines the samples from a set of nested sampling runs with differing
+#' numbers of live points into a single run.
+#'
+#' @param x [[ernest_run]]\cr An object containing a nested sampling run.
+#' @param y [[ernest_run]]\cr Another nested sampling run to merge with `x`.
+#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Additional `ernest_run`
+#' objects to merge with `x`. Ignored if `y` is used.
+#' @param .keep [[character(1)]] Strategy for merging the live set of the runs.
+#' * `"first"`: The live set begins after the worst live point appears in the
+#' run. This is the default, and is the most straightforward way to merge runs
+#' with different numbers of live points.
+#' * `"all"`: The live set gets smaller as points die throughout the run. This
+#' is more complicated, but can preserve more information about the live set
+#' if one run ends far earlier than another.
+#'
+#' @returns [[ernest_run]] An object containing the merged nested sampling run.
+#' The `id` field of the [[ernest_rcrd]] component is updated to ensure IDs
+#' range from `1` to the total number of points across all runs.
+#'
+#' @export
+merge.ernest_run <- function(x, y = NULL, ..., .keep = c("first", "all")) {
+  y_arg <- if (!is.null(y)) caller_arg(y) else "`...`"
+  y <- if (is.null(y)) {
+    list2(...)
+  } else {
+    c(list(y), list2(...))
+  }
+
+  # Check that all runs have the expected type and shape
+  for (yi in y) {
+    check_class(yi, "ernest_run", arg = y_arg)
+    if (attr(yi$rcrd, "nvar") != attr(x$rcrd, "nvar")) {
+      cli::cli_abort(
+        "`{y_arg}` must have the same number of variables as `{caller_arg(x)}`."
+      )
+    }
+  }
+
+  # Merge objects together
+  merged_rcrd <- x$rcrd
+  nlive <- 0L
+  for (yi in y) {
+    m <- merge_rcrd(merged_rcrd, yi$rcrd, keep = .keep)
+    nlive <- attr(m, "nlive")
+    merged_rcrd <- m
+  }
+
+  # Update the sampler
+  old_nlive <- x$nlive
+  x$first_update <- as.integer((x$first_update / old_nlive) * nlive)
+  x$update_interval <- as.integer((x$update_interval / old_nlive) * nlive)
+  x$nlive <- nlive
+
+  new_ernest_run(x, merged_rcrd)
+}
+
+merge_rcrd <- function(
+  x,
+  y,
+  keep = c("first", "all"),
+  invalid_run = c("error", "warn", "quiet")
+) {
+  keep <- arg_match(keep)
+  invalid_run <- arg_match(invalid_run)
+  # Reindex the IDs of each group
+  x_ids <- vctrs::vec_group_id(field(x, "id"))
+  y_ids <- vctrs::vec_group_id(field(y, "id"))
+  nlive <- attr(x_ids, "n") + attr(y_ids, "n")
+  y_ids <- y_ids + attr(x_ids, "n")
+  attributes(x_ids) <- NULL
+  attributes(y_ids) <- NULL
+  vctrs::field(x, "id") <- x_ids
+  vctrs::field(y, "id") <- y_ids
+
+  # Merge the points together and sort them
+  out <- sort(c(x, y))
+  out <- if (keep == "first") {
+    merge_rcrd_first(out, nlive)
+  } else {
[…]
