### 2026-05-15 d3afe7d
Implement merge functionality for ernest_run objects, enhance validation checks for ernest_rcrd, remove S3 generic for new_ernest_run.
