+++ NAMESPACE
@@ -57,6 +57,7 @@ S3method(weights,ernest_run)
 export(as_draws)
 export(as_draws_matrix)
 export(as_draws_rvars)
+export(bootstraps)
 export(calculate)
 export(compile)
 export(create_likelihood)

[…]
