### 2026-05-19 de10f13
Simplified ernest_likelihood class with no use of NSE.
