+++ tests/testthat/_snaps/likelihood-class.md
@@ -3,25 +3,27 @@
     Code
       ll
     Message
-      Scalar Log-likelihood Function
-      function (x) 
-      {
-          x <- matrix(x, ncol = length(x))
-          distval <- stats::mahalanobis(x, center = mean, cov = sigma)
-          exp(-(3 * log(2 * pi) + logdet + distval)/2)
-      }
+      <ernest_likelihood>
+      <function(x) {
+        x <- matrix(x, ncol = length(x))
+        distval <- stats::mahalanobis(x, center = mean, cov = sigma)
+        exp(-(3 * log(2 * pi) + logdet + distval) / 2)
+      }>
+      v Interface: "scalar_fn"
+      v Non-finite handling: "warn"
 
 # ernest_likelihood / produces likelihood from `vectorized_fn`
 
     Code
       mat_ll
     Message
-      Vectorized Log-likelihood Function
-      function (x) 
-      {
-          distval <- stats::mahalanobis(x, center = mean, cov = sigma)
-          exp(matrix(-(3 * log(2 * pi) + logdet + distval)/2, nrow = nrow(x)))
-      }
+      <ernest_likelihood>
+      <function(x) {
+        distval <- stats::mahalanobis(x, center = mean, cov = sigma)
+        exp(matrix(-(3 * log(2 * pi) + logdet + distval) / 2, nrow = nrow(x)))
+      }>
+      v Interface: "vectorized_fn"
+      v Non-finite handling: "warn"
 
 # Missing values in the log-likelihood
 
@@ -34,10 +36,7 @@
       Caused by error in `compile()`:
       ! Error when creating the live set.
       Caused by error:
-      ! Couldn't calculate the log-lik of #.#, #.#, -#.#, #.#, #.#, #.#, #.#, -#.#, #.#, #.#, -#.#, #.#, #.#, -#.#, -#.#, #.#, #.#, -#.#, ..., -#.#, and -#.#.
-      Caused by error:
-      ! log-lik. values must be either finite or `-Inf`.
-      x Detected non-viable value: `NA`.
+      ! `log_lik` cannot return NA.
 
 ---
 
@@ -48,7 +47,7 @@
       Warning:
       <ernest_sampler> threw a warning during compilation
       Caused by warning:
-      ! Replacing `NA` with `-Inf`.
+      ! Replacing log-lik. values with `-Inf`: NA
     Message
       Nested sampling run specification:
       * No. points: 500
