### 2026-05-20 e8b4cda
Enabled likelihood functions for parallel runs.
