### 2026-05-20 eebd665
Enabled likelihood functions for parallel runs.
