### 2026-05-20 bb912bb
Refactor prior for improved functionality and clarity.
