+++ tests/testthat/_snaps/likelihood-class.md
@@ -52,5 +52,5 @@
       Nested sampling run specification:
       * No. points: 500
       * Sampling method: 25-step random walk sampling (acceptance target = 50%)
-      * Prior: uniform prior distribution with 2 dimensions (A and B)
+      * Prior: <uniform_prior/ernest_prior> (2 dims.)
 
