+++ tests/testthat/_snaps/prior-class.md
@@ -2,15 +2,27 @@
 
     Code
       pr
-    Output
-      custom prior distribution with 3 dimensions (A, B, and C)
+    Message
+      <custom_prior/ernest_prior> (3 dims.)
+      <function(x) {
+        stats::qnorm(x, mean = c(-1, 0, 1))
+      }>
+      v Names: "A", "B", and "C"
+      v Interface: "point_fn"
 
 # ernest_prior / produces prior from `vectorized_fn`
 
     Code
       mat_pr
-    Output
-      custom prior distribution with 3 dimensions (A, B, and C)
+    Message
+      <custom_prior/ernest_prior> (3 dims.)
+      <function(x) {
+        y <- stats::qnorm(c(x), mean = rep(c(-1, 0, 1), each = nrow(x) %||% 1))
+        dim(y) <- dim(x)
+        y
+      }>
+      v Names: "A", "B", and "C"
+      v Interface: "vectorized_fn"
 
 # Zero-length prior fails
 
@@ -25,7 +37,7 @@
     Code
       create_prior(prior_fn, names = LETTERS[1])
     Condition
-      Error:
-      ! Can't convert `prior$fn(x)` <double[,0]> to <double[,1]>.
+      Error in `vectorized_prior()`:
+      ! Can't convert `y` <double[,0]> to <double[,1]>.
       Non-recyclable dimensions.
 
