### 2026-05-20 f22a91f
Add parallel_prior function.
