+++ NAMESPACE
@@ -10,6 +10,7 @@ S3method(calculate,ernest_run)
 S3method(compile,ernest_run)
 S3method(compile,ernest_sampler)
 S3method(format,ernest_lrps)
+S3method(format,ernest_prior)
 S3method(format,ernest_rcrd)
 S3method(format,multi_ellipsoid)
 S3method(format,rwmh_cube)
