+++ DESCRIPTION
@@ -33,6 +33,7 @@ Suggests:
     extraDistr,
     ggdist,
     knitr,
+    mirai,
     mvtnorm,
     patchwork,
     rmarkdown,
