+++ R/parallelization.R
@@ -59,6 +59,144 @@ parallel_prior <- function(
   )
 }
 
+pgenerate <- function(
+  x,
+  workers,
+  info,
+  control,
+  show_progress,
+  call = caller_env()
+) {
+  check_parallel_enabled(x, call)
+  worker_nlive <- if (isTRUE(workers)) {
+    default_worker_nlive(x$nlive, x$lrps$nvar)
+  } else {
+    vec_cast(workers, integer())
+  }
+  if (any(worker_nlive < x$ndim) || vctrs::vec_any_missing(worker_nlive)) {
+    stop_input_type(
+      workers,
+      c("`TRUE`", "a vector of integers greater than or equal to `x$ndim`"),
+      call = call
+    )
+  }
+  tot_nlive <- sum(worker_nlive) %|% 0L
+  if (tot_nlive != x$nlive) {
+    stop_input_type(
+      workers,
+      c("`TRUE`", "a vector of integers summing to `x$nlive`"),
+      call = call
+    )
+  }
+
+  # Break apart the IDS into the appropriate number of workers
+  ids <- vctrs::vec_chop(sample.int(x$nlive), sizes = worker_nlive)
+  split_x <- partition_run(x, ids, info, control)
+  prev_rcrd <- if (is.null(x$rcrd)) {
+    NULL
+  } else {
+    x$rcrd[vctrs::vec_as_location(field(x$rcrd, "neval") != 0L, length(x$rcrd))]
+  }
+
+  # Run the workers in parallel
+  lrps_ <- x$lrps
+  m <- mirai::mirai_map(
+    split_x,
+    \(xi) {
+      cur_env <- list2env(xi[c("unit", "log_lik", "birth_lik")])
+      nested_sampling_impl_(
+        live_env = cur_env,
+        lrps = lrps_,
+        sampler_info = xi$info,
+        control = xi$control,
+        show_progress = FALSE
+      )
+    },
+    nested_sampling_impl_ = nested_sampling_impl,
+    lrps_ = x$lrps
+  )
+
+  opts <- c(".stop", if (show_progress) ".progress" else NULL)
+  m_out <- mirai::collect_mirai(m, options = opts)
+  rcrd <- unpartition_runs(m_out, ids, prev_rcrd)
+  new_ernest_run(x, rcrd)
+}
+
+default_worker_nlive <- function(nlive, ndim, call = caller_env()) {
+  ndaemons <- mirai::info()[["connections"]]
+  nlive_per_worker <- nlive %/% ndaemons
+  if (nlive_per_worker < (ndim + 1)) {
+    cli::cli_warn(
+      "The number of live points per daemon has been set to `ndim + 1`.",
+      call = call
+    )
+    nlive_per_worker <- ndim + 1
+  }
+  workers <- rep(nlive_per_worker, ndaemons)
+  workers[[1]] <- workers[[1]] + (nlive - sum(workers))
+  vec_cast(workers, integer())
+}
+
+#' Split the live set into a series of threads.
+#'
+#' @param x The sampler.
+#' @param List of integer vectors, each containing the IDs for that worker.
+#'
+#' @returns Named list of IDs by worker and environments.
+#' @noRd
+partition_run <- function(x, slices, control, info) {
+  x_rcrd <- x$rcrd %||% NULL
+  split_info <- function(split_nlive) {
+    frac_nlive <- split_nlive / info$nlive
+    list(
+      seed = info$seed,
+      nlive = split_nlive,
+      first_update = as.integer(info$first_update * frac_nlive),
+      update_interval = as.integer(info$update_interval * frac_nlive)
+    )
+  }
+
+  lapply(slices, \(slice) {
+    list(
[…]
