+++ tests/testthat/_snaps/parallelization.md
@@ -1,0 +1,30 @@
+# generate & mirai / can run a parallel sampler
+
+    Code
+      run
+    Message
+      Nested sampling run:
+      * No. points: 500
+      * Sampling method: 25-step random walk sampling (acceptance target = 50%)
+      * Prior: <crated_prior/ernest_prior> (2 dims.)
+      -- Results ---------------------------------------------------------------------
+      * Iterations: 4203
+      * Likelihood evals.: 89732
+      * Log-evidence: -6.0544 (± 0.1297)
+      * Information: 5.087
+
+# generate & mirai / can run a parallel sampler with existing rcrd
+
+    Code
+      run
+    Message
+      Nested sampling run:
+      * No. points: 500
+      * Sampling method: 25-step random walk sampling (acceptance target = 50%)
+      * Prior: <crated_prior/ernest_prior> (2 dims.)
+      -- Results ---------------------------------------------------------------------
+      * Iterations: 4538
+      * Likelihood evals.: 98107
+      * Log-evidence: -6.0543 (± 0.1166)
+      * Information: 5.087
+

[…]
