### 2026-05-22 e749f18
Implement parallel sampling support.
