+++ R/generate.R
@@ -83,6 +83,7 @@ generate.ernest_sampler <- function(
   max_evaluations = NULL,
   min_logz = 0.05,
   show_progress = NULL,
+  parallel = FALSE,
   ...
 ) {
   if (is.null(show_progress)) {
@@ -97,6 +98,15 @@ generate.ernest_sampler <- function(
   x <- compile(x, ...)
   info <- get_sampler_info(x)
 
+  if (!isFALSE(parallel)) {
+    return(pgenerate(
+      x = x,
+      workers = parallel,
+      info = info,
+      control = control,
+      show_progress = show_progress
+    ))
+  }
   results <- nested_sampling_impl(
     live_env = x$live_env,
     lrps = x$lrps,
@@ -118,6 +128,7 @@ generate.ernest_run <- function(
   max_evaluations = NULL,
   min_logz = 0.05,
   show_progress = NULL,
+  parallel = FALSE,
   ...
 ) {
   if (is.null(show_progress)) {
@@ -138,6 +149,15 @@ generate.ernest_run <- function(
   )
   info <- get_sampler_info(x)
 
+  if (!isFALSE(parallel)) {
+    return(pgenerate(
+      x = x,
+      workers = parallel,
+      info = info,
+      control = control,
+      show_progress = show_progress
+    ))
+  }
   results <- nested_sampling_impl(
     live_env = x$live_env,
     lrps = x$lrps,
