### 2026-05-22 c06d0e2
Add glance summary methods for ernest_run and ernest_rcrd.
