### 2026-05-22 c7588f6
Add glance summary methods for ernest_run and ernest_rcrd.
