+++ R/parallel.R
@@ -1,259 +1,0 @@
-#' Generate nested sampling runs in parallel.
-#'
-#' @description
-#' `r lifecycle::badge("experimental")`
-#'
-#' The cost of a nested sampling run depends on the distance between the prior
-#' and posterior distributions and on the number of live points. With
-#' `allow_par = TRUE`, [generate()] uses \CRANpkg{mirai} to split the live set
-#' across workers, run sampling in parallel, then merge results before
-#' estimating model evidence.
-#'
-#' To run in parallel, daemons must be set with [mirai::daemons()]. Otherwise,
-#' [mirai::require_daemons()] throws an error when `allow_par = TRUE`.
-#'
-#' User-supplied [ernest_likelihood] and [ernest_prior] functions must be
-#' self-contained and must not depend on objects in the global environment.
-#' This keeps serialization predictable and avoids sending large, accidental
-#' dependencies to workers.
-#'
-#' @section Creating self-contained functions:
-#'
-#' Self-contained likelihood and prior functions should follow these
-#' guidelines:
-#'
-#' 1. Call package functions with explicit `::` namespaces, e.g.
-#' `extraDistr::qtnorm()`. Alternatively, call `library()` inside the function
-#' if you need to attach a package.
-#'
-#' 2. Declare all data dependencies explicitly, similar to the pattern used in
-#' [create_likelihood].
-#'
-#' 3. Any helper functions (closures) called within a self-contained function
-#' must itself be self-contained.
-#'
-#' @section Setting daemons:
-#'
-#' How and where parallelisation occurs is determined by [mirai::daemons()].
-#' Daemons are persistent background processes that execute parallel
-#' computations locally or across a network.
-#'
-#' Daemons must be set before parallel execution. Otherwise, calling
-#' `generate(..., allow_par = TRUE)` throws an error.
-#'
-#' Usually, daemons are set once per session and can be left running while idle
-#' because they use minimal resources. The following sets up 6 local daemons:
-#'
-#' ```r
-#' mirai::daemons(6)
-#' ```
-#'
-#' Function arguments:
-#'
-#' * `n`: the number of daemons to launch on your local machine, e.g.
-#'   `mirai::daemons(6)`. As a rule of thumb, for maximum efficiency this should
-#'   be (at most) one less than the number of cores on your machine, leaving one
-#'   core for the main R process.
-#' * `url` and `remote`: used to set up and launch daemons for distributed
-#'   computing over the network. See [mirai::daemons()] for more details.
-#'
-#' Daemons persist for the duration of your session. To reset and shut them
-#' down:
-#'
-#' ```r
-#' mirai::daemons(0)
-#' ```
-#'
-#' All daemons automatically terminate when your session ends. You do not need
-#' to explicitly terminate daemons in this instance, although it is still good
-#' practice to do so.
-#'
-#' @references
-#' ernest's parallelisation is powered by \CRANpkg{mirai}. See the
-#' [mirai website](https://mirai.r-lib.org/) for more details.
-#'
-#' \CRANpkg{crate} provides a simple method for creating self-contained
-#' functions. Consult that package for more details.
-#'
-#' This documentation is based on the
-#' [in_parallel](https://purrr.tidyverse.org/reference/in_parallel.html)
-#' function from the \pkg{purrr} package.
-#'
-#' @seealso [generate()] for performing nested sampling runs.
-#'
-#' @examples
[…]
