+++ man/glance.ernest_run.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/run-class.R
+\name{glance.ernest_run}
+\alias{glance.ernest_run}
+\title{One-row glance summary for a nested sampling run}
+\usage{
+\method{glance}{ernest_run}(x, ...)
+}
+\arguments{
+\item{x}{[\link{ernest_run}]\cr A nested sampling run.}
+}
+\value{
+A one-row \code{tibble} with \code{nlive}, \code{nvar}, \code{niter}, \code{neval},
+\code{log_evidence}, \code{log_evidence_err}, \code{information}, and \code{seed}.
+}
+\description{
+Returns a compact tibble with the key run-level diagnostics.
+}
+\examples{
+data(example_run)
+glance(example_run)
+}
+\seealso{
+\code{\link[=summary.ernest_run]{summary.ernest_run()}}
+}
