+++ NAMESPACE
@@ -19,6 +19,8 @@ S3method(format,unif_cube)
 S3method(format,unif_ellipsoid)
 S3method(generate,ernest_run)
 S3method(generate,ernest_sampler)
+S3method(glance,ernest_rcrd)
+S3method(glance,ernest_run)
 S3method(merge,ernest_run)
 S3method(plot,ernest_estimate)
 S3method(plot,ernest_run)
@@ -66,6 +68,7 @@ export(create_prior)
 export(create_uniform_prior)
 export(ernest_sampler)
 export(generate)
+export(glance)
 export(multi_ellipsoid)
 export(new_ernest_lrps)
 export(propose)
@@ -84,6 +87,7 @@ importFrom(cli,symbol)
 importFrom(generics,calculate)
 importFrom(generics,compile)
 importFrom(generics,generate)
+importFrom(generics,glance)
 importFrom(generics,visualize)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,autoplot)
