### 2026-05-25 d4559fc
Introduced documentation for parallelization features and usage.
