+++ R/parallelization.R
@@ -1,3 +1,65 @@
+#' Parallelization in ernest
+#'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#'
+#' Nested sampling runs can be performed in parallel using \CRANpkg{mirai}.
+#'
+#' To parallelize a run:
+#' - Create a portable likelihood with `parallel_likelihood()`.
+#' - Create a portable prior with `parallel_prior()` or use one of ernest's
+#' [special priors][special_priors].
+#' - Set worker daemons with `mirai::daemons()` and ensure workers can load
+#' ernest.
+#' - Call [`generate()`][generate-ernest] with `parallel` != `FALSE`.
+#'
+#' @param scalar_fn,point_fn,vectorized_fn `[function]`\cr Used identically to
+#' [create_likelihood()] or [create_prior()], but with the additional
+#' requirement that the functions are 'fresh', meaning they should be declared
+#' in the call to `parallel_likelihood()` or `parallel_prior()` rather than
+#' stored in a variable.
+#' @param ... Named arguments that should be captured in the function's
+#' environment and serialized for remote execution.
+#' @inheritParams create_prior
+#' @inheritParams create_likelihood
+#'
+#' @returns
+#' For `parallel_likelihood()`, a [carrier::crate] with the additional class
+#' [ernest_likelihood].
+#'
+#' For `parallel_prior()`, an [ernest_prior] whose transformation function is
+#' a [carrier::crate].
+#'
+#' @section Parallelized nested sampling:
+#' Splitting an initial live set into subruns lets you use many live points
+#' without running a single very large job. Run several smaller nested sampling
+#' jobs concurrently and merge their records to obtain the same statistical
+#' benefits as a single large run.
+#'
+#' When [`generate()`][generate-ernest] is called with `parallel != FALSE`, the resulting run
+#' contains a `parallel` element with per-worker summaries (see
+#' [glance.ernest_run()]) useful for diagnostics.
+#'
+#' @section Daemons:
+#' How parallelization occurs is determined by [mirai::daemons()]. Daemons must
+#' be set prior to calling `generate()` with `parallel != FALSE`, otherwise an
+#' error will be thrown.
+#'
+#' It is usual to set daemons once per session. You can leave them running on
+#' your local machine as they consume almost no resources whilst waiting to
+#' receive tasks. The following sets up 6 daemons locally:
+#'
+#' ```r
+#' mirai::daemons(6)
+#' ```
+#'
+#' @seealso [mirai::daemons()] [carrier::crate()] [generate-ernest]
+#'
+#' @references Documentation adapted from the \CRANpkg{mirai},
+#' \CRANpkg{carrier}, and \CRANpkg{purrr} packages.
+#'
+#' @rdname parallelization
+#' @export
 parallel_likelihood <- function(
   scalar_fn,
   vectorized_fn,
@@ -22,6 +84,8 @@ parallel_likelihood <- function(
   )
 }
 
+#' @rdname parallelization
+#' @export
 parallel_prior <- function(
   point_fn,
   vectorized_fn,
@@ -59,6 +123,20 @@ parallel_prior <- function(
   )
 }
 
+#' Perform parallel nested sampling
+#'
+#' @param x An ernest_sampler/ernest_run object.
+#' @param workers Either `TRUE` to automatically use all available daemons,
+#' or a vector of integers specifying the number of live points to allocate to
+#' each daemon.
+#' @param info,control Arguments controlling the parent run.
[…]
