+++ NAMESPACE
@@ -71,6 +71,8 @@ export(generate)
 export(glance)
 export(multi_ellipsoid)
 export(new_ernest_lrps)
+export(parallel_likelihood)
+export(parallel_prior)
 export(propose)
 export(rwmh_cube)
 export(slice_rectangle)

[…]
