+++ tests/testthat/test-parallelization.R
@@ -1,84 +1,107 @@
-parallel_lik <- NULL
-parallel_pt <- NULL
-
-test_that("parallel ernest_likelihood works", {
-  parallel_lik <<- parallel_likelihood(
-    vectorized_fn = function(x) {
-      sigma <- 0.1
-      mu1 <- c(1, 1)
-      mu2 <- -c(1, 1)
-      sigma_inv <- diag(2) / 0.1**2
-
-      dx1 <- -0.5 * stats::mahalanobis(x, c(1, 1), sigma_inv, inverted = TRUE)
-      dx2 <- -0.5 * stats::mahalanobis(x, c(-1, -1), sigma_inv, inverted = TRUE)
-      matrixStats::colLogSumExps(rbind(dx1, dx2))
+test_that("Crated func. envs. are attached to the search path", {
+  # Median and qnorm are both in the stats package
+  parallel_lik <- parallel_likelihood(
+    function(x) {
+      median(x)
     }
   )
-  expect_s3_class(parallel_lik, c("ernest_likelihood", "crate"))
-})
+  expect_equal(parallel_lik(c(1, 2, 3)), 2)
 
-test_that("parallel ernest_prior works", {
-  parallel_pt <<- parallel_prior(
-    function(x) {
-      (x * 10) - 5
-    },
-    .names = c("a", "b")
+  parallel_v_prior <- parallel_prior(
+    vectorized_fn = function(x) qnorm(x),
+    .names = c("A", "B", "C")
   )
-  expect_s3_class(parallel_pt, c("crated_prior", "ernest_prior"))
-  expect_s3_class(attr(parallel_pt, "body"), "crate")
+  expect_equal(
+    parallel_v_prior$fn(matrix(c(0.1, 0.5, 0.9), nrow = 1)),
+    matrix(stats::qnorm(c(0.1, 0.5, 0.9)), nrow = 1)
+  )
 })
 
-describe("pgenerate", {
-  it("fails when log_lik is not a crate", {
-    sampler <- ernest_sampler(\(x) sum(x), parallel_pt)
-    expect_error(
-      generate(sampler, parallel = TRUE),
-      "portable `log_lik` function"
-    )
-  })
+# Portable version of the two Gaussian blobs test likelihood
+parallel_lik <- parallel_likelihood(
+  vectorized_fn = function(x) {
+    dx1 <- -0.5 * mahalanobis(x, mu1, sigma_inv, inverted = TRUE)
+    dx2 <- -0.5 * mahalanobis(x, mu2, sigma_inv, inverted = TRUE)
+    matrixStats::colLogSumExps(rbind(dx1, dx2))
+  },
+  sigma = 0.1,
+  mu1 = c(1, 1),
+  mu2 = -c(1, 1),
+  sigma_inv = diag(2) / 0.1**2
+)
 
-  it("fails when prior is not a crate", {
-    sampler <- ernest_sampler(
-      parallel_lik,
-      create_prior(\(x) cumsum(x), names = LETTERS[1:2])
-    )
-    expect_error(
-      generate(sampler, parallel = TRUE),
-      "portable `prior` function"
-    )
-  })
+test_that("parallelization checks for portable functions and daemons", {
+  expect_error(
+    generate(
+      ernest_sampler(gaussian_blobs$log_lik, gaussian_blobs$prior),
+      parallel = TRUE
+    ),
+    "portable `log_lik` function"
+  )
 
-  it("fails when daemons are not set", {
-    sampler <- ernest_sampler(parallel_lik, parallel_pt)
-    expect_error(
-      generate(sampler, parallel = TRUE),
-      "No daemons set."
-    )
-  })
+  expect_error(
+    generate(
+      ernest_sampler(
+        parallel_lik,
[…]
