### 2026-05-25 dcc1d12
`ernest_rcrd` now stores `id` as a character, allowing for safer merging (though now merging and parallelization is broken)
