+++ R/rcrd-class.R
@@ -5,26 +5,21 @@
 #' replacements in the live set and each point's contribution to evidence
 #' estimation.
 #'
-#' @section Fields:
-#' \describe{
-#' \item{`unit`}{`[double(nvar)]` The sample's coordinates within the unit
-#' hypercube (prior to transformation via the prior distribution).}
-#' \item{`id`}{`[integer(1)]` A unique identifier for each sample in the
-#' sequence of live points.}
-#' \item{`nlive`}{`[integer(1)]` The number of effective live points remaining
-#' when the sample was removed from the live set. Generally corresponds to the
-#' `nlive` of the `ernest_run`, but can be lower if the sample found within a
-#' plateau in the log-likelihood function (see References) or if the sample is
-#' still in the live set at termination.}
-#' \item{`neval`}{`[integer(1)]` The number of likelihood evaluations
-#' required by the sampler to generate this replacement point, reflecting the
-#' difficulty of sampling from the likelihood-restricted prior. Equal to `0`
-#' for points remaining in the live set at termination.}
-#' \item{`log_lik`}{`[double(1)]` The log-likelihood at the sample.}
-#' \item{`birth_lik`}{`[double(1)]` The log-likelihood threshold that this
-#' sample exceeded; the minimum likelihood constraint applied when generating
-#' it. Equals `-Inf` for samples in the initial live set.}
-#' }
+#' @param unit `[matrix()]`\cr Coordinates in the unit hypercube, with one row
+#' per sample and one column per variable.
+#' @param log_lik `[double()]`\cr Log-likelihood values for each sample.
+#' @param id `[character()]`\cr Unique identifier for each point within a
+#' nested sampler.
+#' @param nlive `[integer()]`\cr The number of live points in the run at the
+#' time each sample was generated.
+#' @param neval `[integer()]`\cr The number of likelihood evaluations performed
+#' to generate each sample.
+#' @param birth_lik `[double()]`\cr The log-likelihood threshold at which
+#' each sample was generated, i.e. the log-likelihood of the worst live point
+#' at the time of replacement.
+#'
+#' @returns An `ernest_rcrd` object, which is a vctrs record class designed to
+#' store the history of samples generated during a nested sampling run.
 #'
 #' @references Fowlie, A., Handley, W., Su, L., Nested Sampling with Plateaus,
 #' Monthly Notices of the Royal Astronomical Society, 503(1), 1199–1205,
@@ -39,7 +34,32 @@
 #'
 #' @name ernest_rcrd
 #' @keywords internal
-NULL
+ernest_rcrd <- function(
+  unit = matrix(double(0)),
+  log_lik = double(0),
+  id = character(0),
+  nlive = integer(0),
+  neval = integer(0),
+  birth_lik = double(0)
+) {
+  nvar <- ncol(unit) %||% 0L
+  unit <- vec_cast(unit, to = matrix(double(), ncol = nvar))
+  log_lik <- vec_cast(log_lik, to = double())
+  id <- vec_cast(id, to = character())
+  nlive <- vec_cast(nlive, to = integer())
+  neval <- vec_cast(neval, to = integer())
+  birth_lik <- vec_cast(birth_lik, to = double())
+  params <- vctrs::vec_recycle_common(
+    unit,
+    log_lik,
+    id,
+    nlive,
[…]
