### 2026-05-27 5343a20
Revamped merge method.
