+++ R/merge.R
@@ -7,14 +7,6 @@
 #' @param y [[ernest_run]]\cr Another nested sampling run to merge with `x`.
 #' @param suffix `[[character(2)]]` Suffixes to append to the IDs of `x` and `y`
 #' if there are any duplicate IDs.
-#' @param keep `[[character(1)]]` Specifies what live points to retain from
-#' merging runs together:
-#' * `"first"`: The live set begins after the worst live point appears in the
-#' run. This is the default, and is the most straightforward way to merge runs
-#' with different numbers of live points.
-#' * `"all"`: The live set gets smaller as points die throughout the run. This
-#' is more complicated, but can preserve more information about the live set
-#' if one run ends far earlier than another.
 #' @inheritParams rlang::check_dots_empty
 #'
 #' @returns [[ernest_run]] An object containing the merged nested sampling run.
@@ -24,15 +16,20 @@ merge.ernest_run <- function(
   x,
   y,
   suffix = c(".x", ".y"),
-  keep = c("first", "all"),
   ...
 ) {
   check_class(y, "ernest_run")
+  if (isTRUE(all.equal(x, y))) {
+    cli::cli_abort(
+      "`{caller_arg(x)}` and `{caller_arg(y)}` cannot be identical."
+    )
+  }
   check_dots_empty()
 
   # Merge objects together
-  merged_rcrd <- nlive <- NULL
-  c(merged_rcrd, nlive) %<-% merge_rcrd(x$rcrd, y$rcrd, keep = keep)
+  merged <- merge_rcrd(x$rcrd, y$rcrd, suffix = suffix)
+  merged_rcrd <- merged$rcrd
+  nlive <- merged$nlive
 
   # Update the sampler
   old_nlive <- x$nlive
@@ -42,15 +39,24 @@ merge.ernest_run <- function(
   new_ernest_run(x, merged_rcrd)
 }
 
+#' Merge two `ernest_rcrd` objects together.
+#'
+#' @params x,y ernest rcrd objects to merge together.
+#' @param suffix Suffixes to append to the IDs of `x` and `y`
+#' if there are any duplicate IDs.
+#' @param invalid_run Action to take if the merged rcrd fails validation with
+#' `rcrd_is_run()`. One of `"error"`, `"warn", or `"quiet"`.
+#'
+#' @returns A list with two elements: `rcrd`, the merged `ernest_rcrd` object,
+#' and `nlive`, the number of live points in the merged run.
+#' @noRd
 merge_rcrd <- function(
   x,
   y,
   suffix = c(".x", ".y"),
-  keep = c("first", "all"),
   invalid_run = c("error", "warn", "quiet")
 ) {
   suffix <- vec_cast(suffix, character(2))
-  keep <- arg_match(keep)
   invalid_run <- arg_match(invalid_run)
   # Reindex the IDs of each group
   x_ids <- field(x, "id")
@@ -66,66 +72,49 @@ merge_rcrd <- function(
   vctrs::field(x, "id") <- x_ids
   vctrs::field(y, "id") <- y_ids
 
-  # Merge the points together and sort them
-  out <- sort(c(x, y))
-  out <- if (keep == "first") {
-    merge_rcrd_first(out, nlive)
-  } else {
-    merge_rcrd_all(out, nlive)
-  }
+  out <- sort(vctrs::vec_c(x, y))
+  id_loc <- vctrs::vec_group_loc(field(out, "id"))
+  first_live <- min(vapply(
+    id_loc$loc,
+    function(idx) idx[[length(idx)]],
+    integer(1)
+  ))
+
[…]
