### 2026-05-27 c1fc9c7
Improved merging logic to include summaries of both runs in the merged object.
