### 2026-05-28 f7820d7
Refactor merging logic to use compile_merged_rcrd for improved ID handling.
