+++ tests/testthat/test-merge.R
@@ -55,6 +55,6 @@ test_that("merge_rcrd errors when suffixes still produce duplicate ids", {
   # using identical suffixes for both sides should trigger the "must be unique" error
   expect_error(
     merge_rcrd(x, y, suffix = c(".same", ".same")),
-    "IDs of `x` and `y` must be unique"
+    "must contain 20 unique IDs"
   )
 })
