### 2026-05-28 d8ed2b6
Flattened parallelization implementation.
