+++ R/parallelization.R
@@ -139,134 +139,108 @@ crate_fn <- function(
   )
 }
 
-#' Perform parallel nested sampling
+#' Run nested sampling in parallel across multiple daemons
 #'
-#' @param x An ernest_sampler/ernest_run object.
-#' @param workers Either `TRUE` to automatically use all available daemons,
-#' or a vector of integers specifying the number of live points to allocate to
-#' each daemon.
-#' @param info,control Arguments controlling the parent run.
-#' @param show_progress If `TRUE`, a progress bar will be displayed for the
-#' parallel execution.
+#' @param x The ernest_sampler object
+#' @param sampler_info A list containing information about the sampler.
+#' @param control parameters for the nested sampling run, generated from
+#' `set_run_control()`.
+#' @param show_progress Logical. If `TRUE`, displays a progress bar during
+#' sampling.
+#' @param parallel Logical or integer. The number of parallel workers to use,
+#' or `TRUE` to use all available daemons.
 #'
-#' @returns Same as `generate()`, but with the `parallel` element of the output
-#' containing per-worker glance summaries.
-#'
+#' @returns The output of `nested_sampling_impl()`, plus a `parallel` element
+#' containing per-worker glances.
 #' @noRd
-pgenerate <- function(
+nested_sampling_parallel <- function(
   x,
-  workers,
-  info,
+  sampler_info,
   control,
   show_progress,
-  call = caller_env()
+  parallel
 ) {
-  check_parallel_enabled(x, call)
-  daemon_nlive <- if (isTRUE(workers)) {
-    default_daemon_nlive(x$nlive, x$lrps$nvar, call = call)
+  check_parallel_enabled(x, call = caller_env())
+  parallel <- if (isTRUE(parallel)) {
+    mirai::info()[["connections"]]
   } else {
-    vctrs::vec_cast(workers, integer())
+    parallel
   }
-  nvar <- as.integer(x$lrps$nvar)
-  if (vctrs::vec_any_missing(daemon_nlive) || any(daemon_nlive < (nvar + 1L))) {
-    stop_input_type(
-      workers,
-      c("`TRUE`/`FALSE`", "a vector of integers each >= nvar + 1"),
-      call = call
-    )
-  }
-  tot_nlive <- sum(daemon_nlive) %|% 0L
-  if (tot_nlive != x$nlive) {
-    stop_input_type(
-      workers,
-      c("`TRUE`/`FALSE`", "a vector of integers summing to `x$nlive`"),
-      call = call
-    )
-  }
+  live_env <- x$live_env
+  ids <- env_get(live_env, "id")
+  nlive <- x$nlive
 
-  # Break apart the IDS into the appropriate number of workers
-  ids <- vctrs::vec_chop(sample.int(x$nlive), sizes = daemon_nlive)
-  split_x <- partition_run(x, ids, control, info)
-  prev_rcrd <- if (is.null(x$rcrd)) {
-    NULL
-  } else {
-    x$rcrd[vctrs::vec_as_location(field(x$rcrd, "neval") != 0L, length(x$rcrd))]
-  }
+  ids_by_daemon <- allocate_nlive(ids, parallel)
+  parallel_runs <- partition_run(
+    live_env,
+    ids_by_daemon,
+    control,
+    sampler_info,
+    x$rcrd
+  )
 
-  # Run the workers in parallel
-  lrps_ <- x$lrps
+  # Run runs in parallel
   m <- mirai::mirai_map(
-    split_x,
-    \(xi) {
-      cur_env <- list2env(xi[c("unit", "log_lik", "birth_lik")])
+    parallel_runs,
+    \(run) {
[…]

[…]
