### 2026-05-28 5416c1e
Refactor control parameter handling in generate() functions.
