+++ R/generate.R
@@ -95,18 +95,20 @@ generate.ernest_sampler <- function(
     show_progress <- getOption("rlib_message_verbosity", "default") != "quiet"
   }
   check_bool(show_progress)
-  control <- new_generate_control(
+  control <- generate_control(
     max_iterations,
     max_evaluations,
-    min_logz
+    min_logz,
+    seed = attr(x, "seed"),
+    nlive = x$nlive,
+    first_update = x$first_update,
+    update_interval = x$update_interval
   )
   x <- compile(x, ...)
-  info <- get_sampler_info(x)
 
   if (!isFALSE(parallel)) {
     results <- nested_sampling_parallel(
       x,
-      sampler_info = info,
       control = control,
       show_progress = show_progress,
       parallel = parallel
@@ -116,7 +118,6 @@ generate.ernest_sampler <- function(
     results <- nested_sampling_impl(
       live_env = x$live_env,
       lrps = x$lrps,
-      sampler_info = info,
       control = control,
       show_progress = show_progress
     )
@@ -148,19 +149,21 @@ generate.ernest_run <- function(
   }
 
   x_rcrd <- x$rcrd
-  control <- new_generate_control(
+  control <- generate_control(
     max_iterations,
     max_evaluations,
     min_logz,
-    prev_run = x_rcrd
+    seed = attr(x, "seed"),
+    nlive = x$nlive,
+    first_update = x$first_update,
+    update_interval = x$update_interval,
+    rcrd = x_rcrd
   )
-  info <- get_sampler_info(x)
   prev_run <- x$rcrd[get_dead_idx(x$rcrd)]
 
   if (!isFALSE(parallel)) {
     results <- nested_sampling_parallel(
       x,
-      sampler_info = info,
       control = control,
       show_progress = show_progress,
       parallel = parallel
@@ -174,7 +177,6 @@ generate.ernest_run <- function(
     results <- nested_sampling_impl(
       live_env = x$live_env,
       lrps = x$lrps,
-      sampler_info = info,
       control = control,
       show_progress = show_progress
     )
@@ -182,29 +184,68 @@ generate.ernest_run <- function(
   }
 }
 
-#' Generate and validate stopping criteria.
+#' Generate a list of control parameters for nested sampling
 #'
-#' @param max_iterations Maximum number of iterations to perform.
-#' @param max_evaluations Maximum number of likelihood function neval.
-#' @param min_logz Minimum log-ratio between current estimated evidence and
-#' remaining evidence.
-#' @param prev_run The points from a previous run; NULL if no points yet
-#' exist.
-#' @param call Environment for error reporting.
+#' @param max_iterations,max_evaluations,min_logz User-requested stopping
+#' parameters.
+#' @param seed,nlive,first_update,update_interval Parameters from the sampler
+#' object.
+#' @param rcrd An optional `ernest_rcrd` object from a previous run, used to
+#' initialize the control parameters if continuing a run.
 #'
-#' @return A named list containing `max_iterations`, `max_evaluations`,
-#' `min_logz`, `last_criterion`, `log_z`, `log_vol`, `cur_iter`, and `cur_eval`.
+#' @return A named list containing:
+#' * Run meta info: `seed`, `nlive`, `first_update`, `update_interval`
+#' * Validated stopping criteria: `max_iterations`, `max_evaluations`,
+#' `min_logz`.
+#' * Run state: `last_criterion`, `log_z`, `log_vol`, `cur_iter`, `cur_eval`.
 #' @noRd
-new_generate_control <- function(
[…]
