### 2026-05-28 3632076
Introduced a `parallel` parameter in `generate_control` to control nlive setting, replaced `check_unique_names` with a vctrs function.
