### 2026-05-29 5d291a0
Refactor slice sampling implementation to introduce adaptive steps.
