+++ tests/testthat/test-slice_rectangle.R
@@ -5,22 +5,20 @@ set.seed(42)
 #' error messages
 test_that("slice can be called by user", {
   default <- slice_rectangle()
-  expect_snapshot(slice_rectangle(enlarge = 0.5), error = TRUE)
-  expect_snapshot(slice_rectangle(enlarge = NA))
-  expect_equal(default$enlarge, 1)
+  expect_snapshot(slice_rectangle(steps = 0), error = TRUE)
+  expect_snapshot(slice_rectangle(adaptive = TRUE, max_steps = 2), error = TRUE)
+  expect_equal(default$steps, 3L)
   expect_snapshot(default)
 })
 
 describe("slice class", {
-  ptypes <- list(lower = double(), upper = double(), n_accept = integer())
   it("Can be built and propose points", {
     obj <- expect_all_proposals(
       new_slice_rectangle,
       unit_log_fn = fn,
-      nvar = 2,
-      extra_args = "rect"
+      nvar = 2
     )
-    expect_lrps(obj, subclass = "slice_rectangle", !!!ptypes)
+    expect_lrps(obj, subclass = "slice_rectangle")
     expect_snapshot(obj)
   })
 
@@ -28,9 +26,10 @@ describe("slice class", {
     obj <- new_slice_rectangle(fn, nvar = 2)
     samples <- run_sampler(obj)
     new_obj <- update_lrps(obj, samples$unit)
-    expect_lrps(new_obj, subclass = "slice_rectangle", !!!ptypes)
-    expect_identical(new_obj$cache$lower, rep(0, 2))
-    expect_identical(new_obj$cache$upper, rep(1, 2))
+
+    expect_true(
+      inherits(new_obj, "slice_rectangle") || inherits(new_obj, "rwmh_cube")
+    )
     new_samples <- run_sampler(new_obj)
 
     skip_extended()
@@ -45,7 +44,6 @@ describe("slice class", {
       obj,
       "slice_rectangle",
       reset = "n_accept",
-      ptypes = ptypes
     )
   })
 })
@@ -55,4 +53,42 @@ test_that("slice_rectangle can provide good results", {
   expect_gaussian_run(slice_rectangle())
   expect_3D_run(slice_rectangle())
   expect_eggbox_run(slice_rectangle())
+})
+
+describe("adaptive slice", {
+  obj <- new_slice_rectangle(
+    unit_log_fn = fn,
+    nvar = 2,
+    adaptive = TRUE,
+    steps = 2
+  )
+  obj$cache$whitening <- diag(2)
+  obj$cache$mean_dist <- 0.52
+  original <- c(0, 0)
+
+  it("returns distances", {
+    # Provide minimal whitening/mean_dist so adaptive branch can run
+    res <- propose.slice_rectangle(obj, original = original, criterion = -Inf)
+    expect_true(!is.null(res$distance))
+    expect_type(res$distance, "double")
+  })
+
+  unit <- matrix(rnorm(200), ncol = 2)
+  for (i in seq(100)) {
+    unit[i, ] <- propose.slice_rectangle(
+      obj,
+      original = unit[i, ],
+      criterion = -Inf
+    )$unit
+  }
+
+  it("can be updated", {
+    # Provide a matrix so the adaptive update branch can run
+    new_obj <- update_lrps(obj, unit = unit)
+
+    expect_s3_class(new_obj, "slice_rectangle")
+    # If we got back a slice_rectangle, ensure steps were adjusted within bounds
+    expect_true(new_obj$steps >= 1L && new_obj$steps <= new_obj$max_steps)
+    expect_true(is.numeric(new_obj$cache$mean_dist))
+  })
 })
