+++ R/slice_rectangle.R
@@ -1,66 +1,55 @@
-#' Generate new points with slice sampling
+#' Slice sampling within a bounding hyperrectangle
 #'
-#' Create new samples for the live set by evolving a current point in the set
-#' through slice sampling within a bounding hyperrectangle, shrinking the
-#' rectangle when proposals are rejected.
+#' Propose new live points by evolving an existing live point using slice
+#' sampling inside a bounding hyperrectangle. The rectangle is shrunk when a
+#' proposal is rejected.
 #'
-#' @param enlarge `[double(1)]`\cr Factor by which to inflate the
-#' hyperrectangle's volume before sampling (see Details). Optional, and must be
-#' greater or equal to 1 if provided; if left `NA`, sampling is initially
-#' bounded by the unit hypercube at each iteration.
+#' @param steps `[integer(1)]` Number of reslicing iterations per proposal.
+#' @param adaptive `[logical(1)]` If TRUE, adapt `steps` using the mean
+#'   Mahalanobis distance among live points.
+#' @param max_steps `[integer(1)]` Maximum `steps` allowed when
+#'   `adaptive = TRUE`.
 #'
-#' @returns `[slice_rectangle]`, a named list that inherits from
-#' [[ernest_lrps]].
+#' @returns `[slice_rectangle]` A list inheriting from `ernest_lrps`.
 #'
 #' @details
-#' The slice LRPS generates proposals by uniformly sampling within a bounding
-#' hyperrectangle that contains regions of the parameter space satisfying the
-#' likelihood criterion. Sampling begins by selecting a known live point
-#' \eqn{\theta} that satisfies the criterion. Each iteration proposes
-#' a new point within this rectangle via uniform sampling and compares it
-#' against the criterion; if rejected, a new hyperrectangle is drawn such that
-#' the proposed point is on its boundary and \eqn{\theta} is in its interior.
-#' This continues until either a valid proposal is found or the rectangle has
-#' shrunk to the point where no further clamping operations can be performed.
+#' Sampling starts from a live point that meets the likelihood criterion. Each
+#' iteration draws a uniform proposal inside the rectangle and accepts it if it
+#' satisfies the criterion. Rejected proposals cause the rectangle to be
+#' shrunk so the rejected point lies on the boundary and the original point
+#' remains inside. The process repeats until a valid proposal is found or the
+#' rectangle can no longer be reduced.
 #'
-#' By default, the hyperrectangle spans the extreme values of the current
-#' live set in each dimension. This may risk excluding valid regions
-#' of the parameter space, particularly where the posterior is multimodal or
-#' highly non-Gaussian. To mitigate this, set `enlarge > 1`, which inflates the
-#' hyperrectagle's volume by the specified factor before sampling. Setting
-#' `enlarge` to `NA` disables this behaviour, instead slicing from the unit
-#' hypercube at each iteration.
+#' Use `steps > 1` to increase exploration. When `adaptive = TRUE`, `steps` is
+#' increased or decreased depending on whether proposed moves are closer or
+#' farther than the mean Mahalanobis distance of the live set.
 #'
 #' @references
[…]

[…]
