### 2026-05-29 a06f0bc
Add debugme messages in multi_ellipsoid and slice_rectangle.
