### 2026-05-29 6be041a
Update `unif_ellipsoid` default `enlarge` parameter to 1 and removed warning.
