### 2026-05-29 6f1a411
Update `unif_ellipsoid` default `enlarge` parameter to 1 and removed warning.
