### 2026-06-03 35bdcac
Remove adaptive slice sampling implementation and related parameters from `slice_rectangle`.
