+++ R/slice_rectangle.R
@@ -5,10 +5,6 @@
 #' proposal is rejected.
 #'
 #' @param steps `[integer(1)]` Number of reslicing iterations per proposal.
-#' @param adaptive `[logical(1)]` If TRUE, adapt `steps` using the mean
-#'   Mahalanobis distance among live points.
-#' @param max_steps `[integer(1)]` Maximum `steps` allowed when
-#'   `adaptive = TRUE`.
 #'
 #' @returns `[slice_rectangle]` A list inheriting from `ernest_lrps`.
 #'
@@ -20,9 +16,11 @@
 #' remains inside. The process repeats until a valid proposal is found or the
 #' rectangle can no longer be reduced.
 #'
-#' Use `steps > 1` to increase exploration. When `adaptive = TRUE`, `steps` is
-#' increased or decreased depending on whether proposed moves are closer or
-#' farther than the mean Mahalanobis distance of the live set.
+#' Use `steps` to increase how many times a point goes through the slicing
+#' procedure before being returned as a sample for the live set. One step
+#' consists of one or more slicing operations, and each slice proceeds until
+#' a point is found that satisfies the likelihood criterion.
+#'
 #'
 #' @references
 #' Neal, R. M. (2000). Slice Sampling (Version 1). arXiv.
@@ -38,11 +36,9 @@
 #' @family ernest_lrps
 #' @export
 slice_rectangle <- function(
-  steps = 3L,
-  adaptive = FALSE,
-  max_steps = 100
+  steps = 3L
 ) {
-  new_slice_rectangle(steps = steps, adaptive = adaptive, max_steps = max_steps)
+  new_slice_rectangle(steps = steps)
 }
 
 #' @noRd
@@ -61,9 +57,6 @@ format.slice_rectangle <- function(x, ...) {
 #' @param nvar  Number of dimensions.
 #' @param max_loop  Maximum number of proposal attempts.
 #' @param steps Number of times to resample from the parameter space.
-#' @param adaptive Whether or not to adapt the number of steps.
-#' @param max_steps Maximum number of steps the sampler can reach when
-#' `adaptive = TRUE`.
 #' @param cache Optional cache environment.
 #' @param call Error info.
 #'
@@ -78,23 +71,11 @@ new_slice_rectangle <- function(
   nvar = NULL,
   max_loop = 1e6L,
   steps = 3L,
-  adaptive = FALSE,
-  max_steps = 100L,
   cache = NULL,
   call = caller_env()
 ) {
   check_number_whole(steps, min = 1, call = call)
-  check_number_whole(max_steps, min = as.double(steps), call = call)
-  check_bool(adaptive, call = call)
   cache <- cache %||% new_environment()
-
-  if ((is_integerish(nvar) && nvar > 0)) {
-    if (adaptive) {
-      env_poke(cache, "distances", vctrs::list_of(.ptype = double()))
-      env_cache(cache, "whitening", NA)
-      env_cache(cache, "mean_dist", NaN)
-    }
-  }
 
   new_ernest_lrps(
     unit_log_fn = unit_log_fn,
@@ -102,8 +83,6 @@ new_slice_rectangle <- function(
     max_loop = max_loop,
     cache = cache,
     steps = as.integer(steps),
-    adaptive = adaptive,
-    max_steps = as.integer(max_steps),
     .class = "slice_rectangle"
   )
 }
@@ -118,26 +97,13 @@ propose.slice_rectangle <- function(
   if (is.null(original)) {
     NextMethod(x)
   } else {
-    res <- if (x$adaptive) {
-      sample <- SliceImplAdaptive(
-        original = original,
[…]
