+++ src/propose-impl.cpp
@@ -48,39 +48,25 @@ cpp11::list RandomWalkImpl(cpp11::doubles original, cpp11::function unit_log_fn,
 cpp11::list SliceImpl(cpp11::doubles original, cpp11::function unit_log_fn,
                       double criterion, int steps, const int max_loop) {
   // Setups
-  vol::Rectangle rect(original.size());
   Eigen::VectorXd next_draw = as_Matrix(original);
   Eigen::VectorXd inner = next_draw;
+  vol::Rectangle rect(original.size());
 
-  size_t draw = 1;
-  for (; draw < max_loop; draw++) {
-    rect.UniformSample(next_draw);
-    double log_lik = unit_log_fn(as_doubles(next_draw));
-    if (log_lik >= criterion) break;
-    if (!rect.Clamp(inner, next_draw)) break;
+  int neval = 0;
+  for (int draw = 0; draw < steps; draw++) {
+    inner = next_draw;
+    rect = vol::Rectangle(original.size());
+    do {
+      rect.UniformSample(next_draw);
+      double log_lik = unit_log_fn(as_doubles(next_draw));
+      neval++;
+      if (log_lik >= criterion) break;
+      if (!rect.Clamp(inner, next_draw)) break;
+    } while (neval < max_loop);
   }
 
   using namespace cpp11::literals;
   return cpp11::writable::list({"unit"_nm = as_doubles(next_draw),
                                 "log_lik"_nm = unit_log_fn(as_doubles(next_draw)),
-                                "neval"_nm = draw});
-}
-
-// Slice sampling with an additional check for the mahalnobis distance
-// between the current point and the previous point.
-[[cpp11::register]]
-cpp11::list SliceImplAdaptive(cpp11::doubles original, cpp11::function unit_log_fn,
-                              double criterion, int steps, const int max_loop,
-                              cpp11::doubles_matrix<> whitening) {
-  cpp11::writable::list sample =
-      SliceImpl(original, unit_log_fn, criterion, steps, max_loop);
-
-  // Mahalanobis distance between original and new point.
-  Eigen::VectorXd diff = as_Matrix(cpp11::as_doubles(sample["unit"]));
-  diff -= as_Matrix(original);
-  double dist = sqrt(diff.transpose() * as_Matrix(whitening) * diff);
-
-  using namespace cpp11::literals;
-  sample.push_back("distance"_nm = dist);
-  return sample;
+                                "neval"_nm = neval});
 }
