### 2026-06-05 cff495f
Enhance loop safety documentation and refactor LRPS constructors for consistency.
