### 2026-06-05 1362f33
Enhance parallelization checks and warnings for `nlive` relative to parameters.
